Crate rusoto_iot

AWS IoT
AWS IoT provides secure, bi-directional communication between Internet-connected devices (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud. You can discover your custom IoT-Data endpoint to communicate with, configure rules for data processing and integration with other services, organize resources associated with each device (Registry), configure logging, and create and manage policies and credentials to authenticate devices.

For more information about how AWS IoT works, see the Developer Guide.

For information about how to use the credentials provider for AWS IoT, see Authorizing Direct Calls to AWS Services.

If you're using the service, you're probably looking for IotClient and Iot.

Structs

AbortConfig
Details of abort criteria to abort the job.

AbortCriteria
Details of abort criteria to define rules to abort the job.

AcceptCertificateTransferRequest
The input for the AcceptCertificateTransfer operation.

Action
Describes the actions associated with a rule.

ActiveViolation
Information about an active Device Defender security profile behavior violation.

AddThingToBillingGroupRequest

AddThingToBillingGroupResponse

AddThingToThingGroupRequest

AddThingToThingGroupResponse

AlertTarget
A structure containing the alert target ARN and the role ARN.

Allowed
Contains information that allowed the authorization.

AssociateTargetsWithJobRequest

AssociateTargetsWithJobResponse

AttachPolicyRequest

AttachPrincipalPolicyRequest
The input for the AttachPrincipalPolicy operation.

AttachSecurityProfileRequest

AttachSecurityProfileResponse

AttachThingPrincipalRequest
The input for the AttachThingPrincipal operation.

AttachThingPrincipalResponse
The output from the AttachThingPrincipal operation.

AttributePayload
The attribute payload.

AuditCheckConfiguration
Which audit checks are enabled and disabled for this account.

AuditCheckDetails
Information about the audit check.

AuditFinding
The findings (results) of the audit.

AuditNotificationTarget
Information about the targets to which audit notifications are sent.

AuditTaskMetadata
The audits that were performed.

AuthInfo
A collection of authorization information.

AuthResult
The authorizer result.

AuthorizerDescription
The authorizer description.

AuthorizerSummary
The authorizer summary.

AwsJobExecutionsRolloutConfig
Configuration for the rollout of OTA updates.

Behavior
A Device Defender security profile behavior.

BehaviorCriteria
The criteria by which the behavior is determined to be normal.

BillingGroupMetadata
Additional information about the billing group.

BillingGroupProperties
The properties of a billing group.

CACertificate
A CA certificate.

CACertificateDescription
Describes a CA certificate.

CancelAuditTaskRequest

CancelAuditTaskResponse

CancelCertificateTransferRequest
The input for the CancelCertificateTransfer operation.

CancelJobExecutionRequest

CancelJobRequest

CancelJobResponse

Certificate
Information about a certificate.

CertificateDescription
Describes a certificate.

CertificateValidity
When the certificate is valid.

ClearDefaultAuthorizerRequest

ClearDefaultAuthorizerResponse

CloudwatchAlarmAction
Describes an action that updates a CloudWatch alarm.

CloudwatchMetricAction
Describes an action that captures a CloudWatch metric.

CodeSigning
Describes the method to use when code signing a file.

CodeSigningCertificateChain
Describes the certificate chain being used when code signing a file.

CodeSigningSignature
Describes the signature for a file.

Configuration
Configuration.

CreateAuthorizerRequest

CreateAuthorizerResponse

CreateBillingGroupRequest

CreateBillingGroupResponse

CreateCertificateFromCsrRequest
The input for the CreateCertificateFromCsr operation.

CreateCertificateFromCsrResponse
The output from the CreateCertificateFromCsr operation.

CreateDynamicThingGroupRequest

CreateDynamicThingGroupResponse

CreateJobRequest

CreateJobResponse

CreateKeysAndCertificateRequest
The input for the CreateKeysAndCertificate operation.

CreateKeysAndCertificateResponse
The output of the CreateKeysAndCertificate operation.

CreateOTAUpdateRequest

CreateOTAUpdateResponse

CreatePolicyRequest
The input for the CreatePolicy operation.

CreatePolicyResponse
The output from the CreatePolicy operation.

CreatePolicyVersionRequest
The input for the CreatePolicyVersion operation.

CreatePolicyVersionResponse
The output of the CreatePolicyVersion operation.

CreateRoleAliasRequest

CreateRoleAliasResponse

CreateScheduledAuditRequest

CreateScheduledAuditResponse

CreateSecurityProfileRequest

CreateSecurityProfileResponse

CreateStreamRequest

CreateStreamResponse

CreateThingGroupRequest

CreateThingGroupResponse

CreateThingRequest
The input for the CreateThing operation.

CreateThingResponse
The output of the CreateThing operation.

CreateThingTypeRequest
The input for the CreateThingType operation.

CreateThingTypeResponse
The output of the CreateThingType operation.

CreateTopicRuleRequest
The input for the CreateTopicRule operation.

CustomCodeSigning
Describes a custom method used to code sign a file.

DeleteAccountAuditConfigurationRequest

DeleteAccountAuditConfigurationResponse

DeleteAuthorizerRequest

DeleteAuthorizerResponse

DeleteBillingGroupRequest

DeleteBillingGroupResponse

DeleteCACertificateRequest
Input for the DeleteCACertificate operation.

DeleteCACertificateResponse
The output for the DeleteCACertificate operation.

DeleteCertificateRequest
The input for the DeleteCertificate operation.

DeleteDynamicThingGroupRequest

DeleteDynamicThingGroupResponse

DeleteJobExecutionRequest

DeleteJobRequest

DeleteOTAUpdateRequest

DeleteOTAUpdateResponse

DeletePolicyRequest
The input for the DeletePolicy operation.

DeletePolicyVersionRequest
The input for the DeletePolicyVersion operation.

DeleteRegistrationCodeRequest
The input for the DeleteRegistrationCode operation.

DeleteRegistrationCodeResponse
The output for the DeleteRegistrationCode operation.

DeleteRoleAliasRequest

DeleteRoleAliasResponse

DeleteScheduledAuditRequest

DeleteScheduledAuditResponse

DeleteSecurityProfileRequest

DeleteSecurityProfileResponse

DeleteStreamRequest

DeleteStreamResponse

DeleteThingGroupRequest

DeleteThingGroupResponse

DeleteThingRequest
The input for the DeleteThing operation.

DeleteThingResponse
The output of the DeleteThing operation.

DeleteThingTypeRequest
The input for the DeleteThingType operation.

DeleteThingTypeResponse
The output for the DeleteThingType operation.

DeleteTopicRuleRequest
The input for the DeleteTopicRule operation.

DeleteV2LoggingLevelRequest

Denied
Contains information that denied the authorization.

DeprecateThingTypeRequest
The input for the DeprecateThingType operation.

DeprecateThingTypeResponse
The output for the DeprecateThingType operation.

DescribeAccountAuditConfigurationRequest

DescribeAccountAuditConfigurationResponse

DescribeAuditTaskRequest

DescribeAuditTaskResponse

DescribeAuthorizerRequest

DescribeAuthorizerResponse

DescribeBillingGroupRequest

DescribeBillingGroupResponse

DescribeCACertificateRequest
The input for the DescribeCACertificate operation.

DescribeCACertificateResponse
The output from the DescribeCACertificate operation.

DescribeCertificateRequest
The input for the DescribeCertificate operation.

DescribeCertificateResponse
The output of the DescribeCertificate operation.

DescribeDefaultAuthorizerRequest

DescribeDefaultAuthorizerResponse

DescribeEndpointRequest
The input for the DescribeEndpoint operation.

DescribeEndpointResponse
The output from the DescribeEndpoint operation.

DescribeEventConfigurationsRequest

DescribeEventConfigurationsResponse

DescribeIndexRequest

DescribeIndexResponse

DescribeJobExecutionRequest

DescribeJobExecutionResponse

DescribeJobRequest

DescribeJobResponse

DescribeRoleAliasRequest

DescribeRoleAliasResponse

DescribeScheduledAuditRequest

DescribeScheduledAuditResponse

DescribeSecurityProfileRequest

DescribeSecurityProfileResponse

DescribeStreamRequest

DescribeStreamResponse

DescribeThingGroupRequest

DescribeThingGroupResponse

DescribeThingRegistrationTaskRequest

DescribeThingRegistrationTaskResponse

DescribeThingRequest
The input for the DescribeThing operation.

DescribeThingResponse
The output from the DescribeThing operation.

DescribeThingTypeRequest
The input for the DescribeThingType operation.

DescribeThingTypeResponse
The output for the DescribeThingType operation.

Destination
Describes the location of the updated firmware.

DetachPolicyRequest

DetachPrincipalPolicyRequest
The input for the DetachPrincipalPolicy operation.

DetachSecurityProfileRequest

DetachSecurityProfileResponse

DetachThingPrincipalRequest
The input for the DetachThingPrincipal operation.

DetachThingPrincipalResponse
The output from the DetachThingPrincipal operation.

DisableTopicRuleRequest
The input for the DisableTopicRuleRequest operation.

DynamoDBAction
Describes an action to write to a DynamoDB table.

The tableName, hashKeyField, and rangeKeyField values must match the values used when you created the table.

The hashKeyValue and rangeKeyvalue fields use a substitution template syntax. These templates provide data at runtime. The syntax is as follows: ${sql-expression}.

You can specify any valid expression in a WHERE or SELECT clause, including JSON properties, comparisons, calculations, and functions. For example, the following field uses the third level of the topic:

"hashKeyValue": "${topic(3)}"

The following field uses the timestamp:

"rangeKeyValue": "${timestamp()}"

DynamoDBv2Action
Describes an action to write to a DynamoDB table.

This DynamoDB action writes each attribute in the message payload into it's own column in the DynamoDB table.

EffectivePolicy
The policy that has the effect on the authorization results.

ElasticsearchAction
Describes an action that writes data to an Amazon Elasticsearch Service domain.

EnableTopicRuleRequest
The input for the EnableTopicRuleRequest operation.

ErrorInfo
Error information.

ExplicitDeny
Information that explicitly denies authorization.

ExponentialRolloutRate
Allows you to create an exponential rate of rollout for a job.

FileLocation
The location of the OTA update.

FirehoseAction
Describes an action that writes data to an Amazon Kinesis Firehose stream.

GetEffectivePoliciesRequest

GetEffectivePoliciesResponse

GetIndexingConfigurationRequest

GetIndexingConfigurationResponse

GetJobDocumentRequest

GetJobDocumentResponse

GetLoggingOptionsRequest
The input for the GetLoggingOptions operation.

GetLoggingOptionsResponse
The output from the GetLoggingOptions operation.

GetOTAUpdateRequest

GetOTAUpdateResponse

GetPolicyRequest
The input for the GetPolicy operation.

GetPolicyResponse
The output from the GetPolicy operation.

GetPolicyVersionRequest
The input for the GetPolicyVersion operation.

GetPolicyVersionResponse
The output from the GetPolicyVersion operation.

GetRegistrationCodeRequest
The input to the GetRegistrationCode operation.

GetRegistrationCodeResponse
The output from the GetRegistrationCode operation.

GetStatisticsRequest

GetStatisticsResponse

GetTopicRuleRequest
The input for the GetTopicRule operation.

GetTopicRuleResponse
The output from the GetTopicRule operation.

GetV2LoggingOptionsRequest

GetV2LoggingOptionsResponse

GroupNameAndArn
The name and ARN of a group.

ImplicitDeny
Information that implicitly denies authorization. When policy doesn't explicitly deny or allow an action on a resource it is considered an implicit deny.

IotAnalyticsAction
Sends messge data to an AWS IoT Analytics channel.

IotClient
A client for the AWS IoT API.

IotEventsAction
Sends an input to an AWS IoT Events detector.

Job
The Job object contains details about a job.

JobExecution
The job execution object represents the execution of a job on a particular device.

JobExecutionStatusDetails
Details of the job execution status.

JobExecutionSummary
The job execution summary.

JobExecutionSummaryForJob
Contains a summary of information about job executions for a specific job.

JobExecutionSummaryForThing
The job execution summary for a thing.

JobExecutionsRolloutConfig
Allows you to create a staged rollout of a job.

JobProcessDetails
The job process details.

JobSummary
The job summary.

KeyPair
Describes a key pair.

KinesisAction
Describes an action to write data to an Amazon Kinesis stream.

LambdaAction
Describes an action to invoke a Lambda function.

ListActiveViolationsRequest

ListActiveViolationsResponse

ListAttachedPoliciesRequest

ListAttachedPoliciesResponse

ListAuditFindingsRequest

ListAuditFindingsResponse

ListAuditTasksRequest

ListAuditTasksResponse

ListAuthorizersRequest

ListAuthorizersResponse

ListBillingGroupsRequest

ListBillingGroupsResponse

ListCACertificatesRequest
Input for the ListCACertificates operation.

ListCACertificatesResponse
The output from the ListCACertificates operation.

ListCertificatesByCARequest
The input to the ListCertificatesByCA operation.

ListCertificatesByCAResponse
The output of the ListCertificatesByCA operation.

ListCertificatesRequest
The input for the ListCertificates operation.

ListCertificatesResponse
The output of the ListCertificates operation.

ListIndicesRequest

ListIndicesResponse

ListJobExecutionsForJobRequest

ListJobExecutionsForJobResponse

ListJobExecutionsForThingRequest

ListJobExecutionsForThingResponse

ListJobsRequest

ListJobsResponse

ListOTAUpdatesRequest

ListOTAUpdatesResponse

ListOutgoingCertificatesRequest
The input to the ListOutgoingCertificates operation.

ListOutgoingCertificatesResponse
The output from the ListOutgoingCertificates operation.

ListPoliciesRequest
The input for the ListPolicies operation.

ListPoliciesResponse
The output from the ListPolicies operation.

ListPolicyPrincipalsRequest
The input for the ListPolicyPrincipals operation.

ListPolicyPrincipalsResponse
The output from the ListPolicyPrincipals operation.

ListPolicyVersionsRequest
The input for the ListPolicyVersions operation.

ListPolicyVersionsResponse
The output from the ListPolicyVersions operation.

ListPrincipalPoliciesRequest
The input for the ListPrincipalPolicies operation.

ListPrincipalPoliciesResponse
The output from the ListPrincipalPolicies operation.

ListPrincipalThingsRequest
The input for the ListPrincipalThings operation.

ListPrincipalThingsResponse
The output from the ListPrincipalThings operation.

ListRoleAliasesRequest

ListRoleAliasesResponse

ListScheduledAuditsRequest

ListScheduledAuditsResponse

ListSecurityProfilesForTargetRequest

ListSecurityProfilesForTargetResponse

ListSecurityProfilesRequest

ListSecurityProfilesResponse

ListStreamsRequest

ListStreamsResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

ListTargetsForPolicyRequest

ListTargetsForPolicyResponse

ListTargetsForSecurityProfileRequest

ListTargetsForSecurityProfileResponse

ListThingGroupsForThingRequest

ListThingGroupsForThingResponse

ListThingGroupsRequest

ListThingGroupsResponse

ListThingPrincipalsRequest
The input for the ListThingPrincipal operation.

ListThingPrincipalsResponse
The output from the ListThingPrincipals operation.

ListThingRegistrationTaskReportsRequest

ListThingRegistrationTaskReportsResponse

ListThingRegistrationTasksRequest

ListThingRegistrationTasksResponse

ListThingTypesRequest
The input for the ListThingTypes operation.

ListThingTypesResponse
The output for the ListThingTypes operation.

ListThingsInBillingGroupRequest

ListThingsInBillingGroupResponse

ListThingsInThingGroupRequest

ListThingsInThingGroupResponse

ListThingsRequest
The input for the ListThings operation.

ListThingsResponse
The output from the ListThings operation.

ListTopicRulesRequest
The input for the ListTopicRules operation.

ListTopicRulesResponse
The output from the ListTopicRules operation.

ListV2LoggingLevelsRequest

ListV2LoggingLevelsResponse

ListViolationEventsRequest

ListViolationEventsResponse

LogTarget
A log target.

LogTargetConfiguration
The target configuration.

LoggingOptionsPayload
Describes the logging options payload.

MetricValue
The value to be compared with the metric.

NonCompliantResource
Information about the resource that was non-compliant with the audit check.

OTAUpdateFile
Describes a file to be associated with an OTA update.

OTAUpdateInfo
Information about an OTA update.

OTAUpdateSummary
An OTA update summary.

OutgoingCertificate
A certificate that has been transferred but not yet accepted.

Policy
Describes an AWS IoT policy.

PolicyVersion
Describes a policy version.

PolicyVersionIdentifier
Information about the version of the policy associated with the resource.

PresignedUrlConfig
Configuration for pre-signed S3 URLs.

PutItemInput
The input for the DynamoActionVS action that specifies the DynamoDB table to which the message data will be written.

RateIncreaseCriteria
Allows you to define a criteria to initiate the increase in rate of rollout for a job.

RegisterCACertificateRequest
The input to the RegisterCACertificate operation.

RegisterCACertificateResponse
The output from the RegisterCACertificateResponse operation.

RegisterCertificateRequest
The input to the RegisterCertificate operation.

RegisterCertificateResponse
The output from the RegisterCertificate operation.

RegisterThingRequest

RegisterThingResponse

RegistrationConfig
The registration configuration.

RejectCertificateTransferRequest
The input for the RejectCertificateTransfer operation.

RelatedResource
Information about a related resource.

RemoveThingFromBillingGroupRequest

RemoveThingFromBillingGroupResponse

RemoveThingFromThingGroupRequest

RemoveThingFromThingGroupResponse

ReplaceTopicRuleRequest
The input for the ReplaceTopicRule operation.

RepublishAction
Describes an action to republish to another topic.

ResourceIdentifier
Information identifying the non-compliant resource.

RoleAliasDescription
Role alias description.

S3Action
Describes an action to write data to an Amazon S3 bucket.

S3Destination
Describes the location of updated firmware in S3.

S3Location
The S3 location.

SalesforceAction
Describes an action to write a message to a Salesforce IoT Cloud Input Stream.

ScheduledAuditMetadata
Information about the scheduled audit.

SearchIndexRequest

SearchIndexResponse

SecurityProfileIdentifier
Identifying information for a Device Defender security profile.

SecurityProfileTarget
A target to which an alert is sent when a security profile behavior is violated.

SecurityProfileTargetMapping
Information about a security profile and the target associated with it.

SetDefaultAuthorizerRequest

SetDefaultAuthorizerResponse

SetDefaultPolicyVersionRequest
The input for the SetDefaultPolicyVersion operation.

SetLoggingOptionsRequest
The input for the SetLoggingOptions operation.

SetV2LoggingLevelRequest

SetV2LoggingOptionsRequest

SigningProfileParameter
Describes the code-signing profile.

SnsAction
Describes an action to publish to an Amazon SNS topic.

SqsAction
Describes an action to publish data to an Amazon SQS queue.

StartOnDemandAuditTaskRequest

StartOnDemandAuditTaskResponse

StartSigningJobParameter
Information required to start a signing job.

StartThingRegistrationTaskRequest

StartThingRegistrationTaskResponse

StatisticalThreshold
A statistical ranking (percentile) which indicates a threshold value by which a behavior is determined to be in compliance or in violation of the behavior.

Statistics
A map of key-value pairs for all supported statistics. Currently, only count is supported.

StepFunctionsAction
Starts execution of a Step Functions state machine.

StopThingRegistrationTaskRequest

StopThingRegistrationTaskResponse

Stream
Describes a group of files that can be streamed.

StreamFile
Represents a file to stream.

StreamInfo
Information about a stream.

StreamSummary
A summary of a stream.

Tag
A set of key/value pairs that are used to manage the resource.

TagResourceRequest

TagResourceResponse

TaskStatistics
Statistics for the checks performed during the audit.

TestAuthorizationRequest

TestAuthorizationResponse

TestInvokeAuthorizerRequest

TestInvokeAuthorizerResponse

ThingAttribute
The properties of the thing, including thing name, thing type name, and a list of thing attributes.

ThingConnectivity
The connectivity status of the thing.

ThingDocument
The thing search index document.

ThingGroupDocument
The thing group search index document.

ThingGroupIndexingConfiguration
Thing group indexing configuration.

ThingGroupMetadata
Thing group metadata.

ThingGroupProperties
Thing group properties.

ThingIndexingConfiguration
The thing indexing configuration. For more information, see Managing Thing Indexing.

ThingTypeDefinition
The definition of the thing type, including thing type name and description.

ThingTypeMetadata
The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when time was deprecated.

ThingTypeProperties
The ThingTypeProperties contains information about the thing type including: a thing type description, and a list of searchable thing attribute names.

TimeoutConfig
Specifies the amount of time each device has to finish its execution of the job. A timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the timer expires, it will be automatically set to TIMED_OUT.

TopicRule
Describes a rule.

TopicRuleListItem
Describes a rule.

TopicRulePayload
Describes a rule.

TransferCertificateRequest
The input for the TransferCertificate operation.

TransferCertificateResponse
The output from the TransferCertificate operation.

TransferData
Data used to transfer a certificate to an AWS account.

UntagResourceRequest

UntagResourceResponse

UpdateAccountAuditConfigurationRequest

UpdateAccountAuditConfigurationResponse

UpdateAuthorizerRequest

UpdateAuthorizerResponse

UpdateBillingGroupRequest

UpdateBillingGroupResponse

UpdateCACertificateRequest
The input to the UpdateCACertificate operation.

UpdateCertificateRequest
The input for the UpdateCertificate operation.

UpdateDynamicThingGroupRequest

UpdateDynamicThingGroupResponse

UpdateEventConfigurationsRequest

UpdateEventConfigurationsResponse

UpdateIndexingConfigurationRequest

UpdateIndexingConfigurationResponse

UpdateJobRequest

UpdateRoleAliasRequest

UpdateRoleAliasResponse

UpdateScheduledAuditRequest

UpdateScheduledAuditResponse

UpdateSecurityProfileRequest

UpdateSecurityProfileResponse

UpdateStreamRequest

UpdateStreamResponse

UpdateThingGroupRequest

UpdateThingGroupResponse

UpdateThingGroupsForThingRequest

UpdateThingGroupsForThingResponse

UpdateThingRequest
The input for the UpdateThing operation.

UpdateThingResponse
The output from the UpdateThing operation.

ValidateSecurityProfileBehaviorsRequest

ValidateSecurityProfileBehaviorsResponse

ValidationError
Information about an error found in a behavior specification.

ViolationEvent
Information about a Device Defender security profile behavior violation.

Enums

AcceptCertificateTransferError
Errors returned by AcceptCertificateTransfer

AddThingToBillingGroupError
Errors returned by AddThingToBillingGroup

AddThingToThingGroupError
Errors returned by AddThingToThingGroup

AssociateTargetsWithJobError
Errors returned by AssociateTargetsWithJob

AttachPolicyError
Errors returned by AttachPolicy

AttachPrincipalPolicyError
Errors returned by AttachPrincipalPolicy

AttachSecurityProfileError
Errors returned by AttachSecurityProfile

AttachThingPrincipalError
Errors returned by AttachThingPrincipal

CancelAuditTaskError
Errors returned by CancelAuditTask

CancelCertificateTransferError
Errors returned by CancelCertificateTransfer

CancelJobError
Errors returned by CancelJob

CancelJobExecutionError
Errors returned by CancelJobExecution

ClearDefaultAuthorizerError
Errors returned by ClearDefaultAuthorizer

CreateAuthorizerError
Errors returned by CreateAuthorizer

CreateBillingGroupError
Errors returned by CreateBillingGroup

CreateCertificateFromCsrError
Errors returned by CreateCertificateFromCsr

CreateDynamicThingGroupError
Errors returned by CreateDynamicThingGroup

CreateJobError
Errors returned by CreateJob

CreateKeysAndCertificateError
Errors returned by CreateKeysAndCertificate

CreateOTAUpdateError
Errors returned by CreateOTAUpdate

CreatePolicyError
Errors returned by CreatePolicy

CreatePolicyVersionError
Errors returned by CreatePolicyVersion

CreateRoleAliasError
Errors returned by CreateRoleAlias

CreateScheduledAuditError
Errors returned by CreateScheduledAudit

CreateSecurityProfileError
Errors returned by CreateSecurityProfile

CreateStreamError
Errors returned by CreateStream

CreateThingError
Errors returned by CreateThing

CreateThingGroupError
Errors returned by CreateThingGroup

CreateThingTypeError
Errors returned by CreateThingType

CreateTopicRuleError
Errors returned by CreateTopicRule

DeleteAccountAuditConfigurationError
Errors returned by DeleteAccountAuditConfiguration

DeleteAuthorizerError
Errors returned by DeleteAuthorizer

DeleteBillingGroupError
Errors returned by DeleteBillingGroup

DeleteCACertificateError
Errors returned by DeleteCACertificate

DeleteCertificateError
Errors returned by DeleteCertificate

DeleteDynamicThingGroupError
Errors returned by DeleteDynamicThingGroup

DeleteJobError
Errors returned by DeleteJob

DeleteJobExecutionError
Errors returned by DeleteJobExecution

DeleteOTAUpdateError
Errors returned by DeleteOTAUpdate

DeletePolicyError
Errors returned by DeletePolicy

DeletePolicyVersionError
Errors returned by DeletePolicyVersion

DeleteRegistrationCodeError
Errors returned by DeleteRegistrationCode

DeleteRoleAliasError
Errors returned by DeleteRoleAlias

DeleteScheduledAuditError
Errors returned by DeleteScheduledAudit

DeleteSecurityProfileError
Errors returned by DeleteSecurityProfile

DeleteStreamError
Errors returned by DeleteStream

DeleteThingError
Errors returned by DeleteThing

DeleteThingGroupError
Errors returned by DeleteThingGroup

DeleteThingTypeError
Errors returned by DeleteThingType

DeleteTopicRuleError
Errors returned by DeleteTopicRule

DeleteV2LoggingLevelError
Errors returned by DeleteV2LoggingLevel

DeprecateThingTypeError
Errors returned by DeprecateThingType

DescribeAccountAuditConfigurationError
Errors returned by DescribeAccountAuditConfiguration

DescribeAuditTaskError
Errors returned by DescribeAuditTask

DescribeAuthorizerError
Errors returned by DescribeAuthorizer

DescribeBillingGroupError
Errors returned by DescribeBillingGroup

DescribeCACertificateError
Errors returned by DescribeCACertificate

DescribeCertificateError
Errors returned by DescribeCertificate

DescribeDefaultAuthorizerError
Errors returned by DescribeDefaultAuthorizer

DescribeEndpointError
Errors returned by DescribeEndpoint

DescribeEventConfigurationsError
Errors returned by DescribeEventConfigurations

DescribeIndexError
Errors returned by DescribeIndex

DescribeJobError
Errors returned by DescribeJob

DescribeJobExecutionError
Errors returned by DescribeJobExecution

DescribeRoleAliasError
Errors returned by DescribeRoleAlias

DescribeScheduledAuditError
Errors returned by DescribeScheduledAudit

DescribeSecurityProfileError
Errors returned by DescribeSecurityProfile

DescribeStreamError
Errors returned by DescribeStream

DescribeThingError
Errors returned by DescribeThing

DescribeThingGroupError
Errors returned by DescribeThingGroup

DescribeThingRegistrationTaskError
Errors returned by DescribeThingRegistrationTask

DescribeThingTypeError
Errors returned by DescribeThingType

DetachPolicyError
Errors returned by DetachPolicy

DetachPrincipalPolicyError
Errors returned by DetachPrincipalPolicy

DetachSecurityProfileError
Errors returned by DetachSecurityProfile

DetachThingPrincipalError
Errors returned by DetachThingPrincipal

DisableTopicRuleError
Errors returned by DisableTopicRule

EnableTopicRuleError
Errors returned by EnableTopicRule

GetEffectivePoliciesError
Errors returned by GetEffectivePolicies

GetIndexingConfigurationError
Errors returned by GetIndexingConfiguration

GetJobDocumentError
Errors returned by GetJobDocument

GetLoggingOptionsError
Errors returned by GetLoggingOptions

GetOTAUpdateError
Errors returned by GetOTAUpdate

GetPolicyError
Errors returned by GetPolicy

GetPolicyVersionError
Errors returned by GetPolicyVersion

GetRegistrationCodeError
Errors returned by GetRegistrationCode

GetStatisticsError
Errors returned by GetStatistics

GetTopicRuleError
Errors returned by GetTopicRule

GetV2LoggingOptionsError
Errors returned by GetV2LoggingOptions

ListActiveViolationsError
Errors returned by ListActiveViolations

ListAttachedPoliciesError
Errors returned by ListAttachedPolicies

ListAuditFindingsError
Errors returned by ListAuditFindings

ListAuditTasksError
Errors returned by ListAuditTasks

ListAuthorizersError
Errors returned by ListAuthorizers

ListBillingGroupsError
Errors returned by ListBillingGroups

ListCACertificatesError
Errors returned by ListCACertificates

ListCertificatesByCAError
Errors returned by ListCertificatesByCA

ListCertificatesError
Errors returned by ListCertificates

ListIndicesError
Errors returned by ListIndices

ListJobExecutionsForJobError
Errors returned by ListJobExecutionsForJob

ListJobExecutionsForThingError
Errors returned by ListJobExecutionsForThing

ListJobsError
Errors returned by ListJobs

ListOTAUpdatesError
Errors returned by ListOTAUpdates

ListOutgoingCertificatesError
Errors returned by ListOutgoingCertificates

ListPoliciesError
Errors returned by ListPolicies

ListPolicyPrincipalsError
Errors returned by ListPolicyPrincipals

ListPolicyVersionsError
Errors returned by ListPolicyVersions

ListPrincipalPoliciesError
Errors returned by ListPrincipalPolicies

ListPrincipalThingsError
Errors returned by ListPrincipalThings

ListRoleAliasesError
Errors returned by ListRoleAliases

ListScheduledAuditsError
Errors returned by ListScheduledAudits

ListSecurityProfilesError
Errors returned by ListSecurityProfiles

ListSecurityProfilesForTargetError
Errors returned by ListSecurityProfilesForTarget

ListStreamsError
Errors returned by ListStreams

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTargetsForPolicyError
Errors returned by ListTargetsForPolicy

ListTargetsForSecurityProfileError
Errors returned by ListTargetsForSecurityProfile

ListThingGroupsError
Errors returned by ListThingGroups

ListThingGroupsForThingError
Errors returned by ListThingGroupsForThing

ListThingPrincipalsError
Errors returned by ListThingPrincipals

ListThingRegistrationTaskReportsError
Errors returned by ListThingRegistrationTaskReports

ListThingRegistrationTasksError
Errors returned by ListThingRegistrationTasks

ListThingTypesError
Errors returned by ListThingTypes

ListThingsError
Errors returned by ListThings

ListThingsInBillingGroupError
Errors returned by ListThingsInBillingGroup

ListThingsInThingGroupError
Errors returned by ListThingsInThingGroup

ListTopicRulesError
Errors returned by ListTopicRules

ListV2LoggingLevelsError
Errors returned by ListV2LoggingLevels

ListViolationEventsError
Errors returned by ListViolationEvents

RegisterCACertificateError
Errors returned by RegisterCACertificate

RegisterCertificateError
Errors returned by RegisterCertificate

RegisterThingError
Errors returned by RegisterThing

RejectCertificateTransferError
Errors returned by RejectCertificateTransfer

RemoveThingFromBillingGroupError
Errors returned by RemoveThingFromBillingGroup

RemoveThingFromThingGroupError
Errors returned by RemoveThingFromThingGroup

ReplaceTopicRuleError
Errors returned by ReplaceTopicRule

SearchIndexError
Errors returned by SearchIndex

SetDefaultAuthorizerError
Errors returned by SetDefaultAuthorizer

SetDefaultPolicyVersionError
Errors returned by SetDefaultPolicyVersion

SetLoggingOptionsError
Errors returned by SetLoggingOptions

SetV2LoggingLevelError
Errors returned by SetV2LoggingLevel

SetV2LoggingOptionsError
Errors returned by SetV2LoggingOptions

StartOnDemandAuditTaskError
Errors returned by StartOnDemandAuditTask

StartThingRegistrationTaskError
Errors returned by StartThingRegistrationTask

StopThingRegistrationTaskError
Errors returned by StopThingRegistrationTask

TagResourceError
Errors returned by TagResource

TestAuthorizationError
Errors returned by TestAuthorization

TestInvokeAuthorizerError
Errors returned by TestInvokeAuthorizer

TransferCertificateError
Errors returned by TransferCertificate

UntagResourceError
Errors returned by UntagResource

UpdateAccountAuditConfigurationError
Errors returned by UpdateAccountAuditConfiguration

UpdateAuthorizerError
Errors returned by UpdateAuthorizer

UpdateBillingGroupError
Errors returned by UpdateBillingGroup

UpdateCACertificateError
Errors returned by UpdateCACertificate

UpdateCertificateError
Errors returned by UpdateCertificate

UpdateDynamicThingGroupError
Errors returned by UpdateDynamicThingGroup

UpdateEventConfigurationsError
Errors returned by UpdateEventConfigurations

UpdateIndexingConfigurationError
Errors returned by UpdateIndexingConfiguration

UpdateJobError
Errors returned by UpdateJob

UpdateRoleAliasError
Errors returned by UpdateRoleAlias

UpdateScheduledAuditError
Errors returned by UpdateScheduledAudit

UpdateSecurityProfileError
Errors returned by UpdateSecurityProfile

UpdateStreamError
Errors returned by UpdateStream

UpdateThingError
Errors returned by UpdateThing

UpdateThingGroupError
Errors returned by UpdateThingGroup

UpdateThingGroupsForThingError
Errors returned by UpdateThingGroupsForThing

ValidateSecurityProfileBehaviorsError
Errors returned by ValidateSecurityProfileBehaviors

Traits

Iot
Trait representing the capabilities of the AWS IoT API. AWS IoT clients implement this trait.