Struct rusoto_iot::DynamoDBAction

pub struct DynamoDBAction {
    pub hash_key_field: String,
    pub hash_key_type: Option<String>,
    pub hash_key_value: String,
    pub operation: Option<String>,
    pub payload_field: Option<String>,
    pub range_key_field: Option<String>,
    pub range_key_type: Option<String>,
    pub range_key_value: Option<String>,
    pub role_arn: String,
    pub table_name: String,
}

Describes an action to write to a DynamoDB table.

The tableName, hashKeyField, and rangeKeyField values must match the values used when you created the table.

The hashKeyValue and rangeKeyvalue fields use a substitution template syntax. These templates provide data at runtime. The syntax is as follows: ${sql-expression}.

You can specify any valid expression in a WHERE or SELECT clause, including JSON properties, comparisons, calculations, and functions. For example, the following field uses the third level of the topic:

"hashKeyValue": "${topic(3)}"

The following field uses the timestamp:

"rangeKeyValue": "${timestamp()}"

Fields
hash_key_field: String

The hash key name.

hash_key_type: Option<String>

The hash key type. Valid values are "STRING" or "NUMBER"

hash_key_value: String

The hash key value.

operation: Option<String>

The type of operation to be performed. This follows the substitution template, so it can be ${operation}, but the substitution must result in one of the following: INSERT, UPDATE, or DELETE.

payload_field: Option<String>

The action payload. This name can be customized.

range_key_field: Option<String>

The range key name.

range_key_type: Option<String>

The range key type. Valid values are "STRING" or "NUMBER"

range_key_value: Option<String>

The range key value.

role_arn: String

The ARN of the IAM role that grants access to the DynamoDB table.

table_name: String

The name of the DynamoDB table.