Struct rusoto_iot::IotClient

pub struct IotClient { /* fields omitted */ }

A client for the AWS IoT API.

Methods

impl IotClient

pub fn new(region: Region) -> IotClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> IotClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,