Struct rusoto_iot::Job

pub struct Job {
    pub comment: Option<String>,
    pub completed_at: Option<f64>,
    pub created_at: Option<f64>,
    pub description: Option<String>,
    pub document_parameters: Option<HashMap<String, String>>,
    pub job_arn: Option<String>,
    pub job_executions_rollout_config: Option<JobExecutionsRolloutConfig>,
    pub job_id: Option<String>,
    pub job_process_details: Option<JobProcessDetails>,
    pub last_updated_at: Option<f64>,
    pub presigned_url_config: Option<PresignedUrlConfig>,
    pub status: Option<String>,
    pub target_selection: Option<String>,
    pub targets: Option<Vec<String>>,
}

The Job object contains details about a job.

Fields
comment: Option<String>

If the job was updated, describes the reason for the update.

completed_at: Option<f64>

The time, in milliseconds since the epoch, when the job was completed.

created_at: Option<f64>

The time, in milliseconds since the epoch, when the job was created.

description: Option<String>

A short text description of the job.

document_parameters: Option<HashMap<String, String>>

The parameters specified for the job document.

job_arn: Option<String>

An ARN identifying the job with format "arn:aws:iot:region:account:job/jobId".

job_executions_rollout_config: Option<JobExecutionsRolloutConfig>

Allows you to create a staged rollout of a job.

job_id: Option<String>

The unique identifier you assigned to this job when it was created.

job_process_details: Option<JobProcessDetails>

Details about the job process.

last_updated_at: Option<f64>

The time, in milliseconds since the epoch, when the job was last updated.

presigned_url_config: Option<PresignedUrlConfig>

Configuration for pre-signed S3 URLs.

status: Option<String>

The status of the job, one of IN_PROGRESS, CANCELED, or COMPLETED.

target_selection: Option<String>

Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a device when the thing representing the device is added to a target group, even after the job was completed by all things originally in the group.

targets: Option<Vec<String>>

A list of IoT things and thing groups to which the job should be sent.