Struct rusoto_iot::TestAuthorizationRequest

pub struct TestAuthorizationRequest {
    pub auth_infos: Vec<AuthInfo>,
    pub client_id: Option<String>,
    pub cognito_identity_pool_id: Option<String>,
    pub policy_names_to_add: Option<Vec<String>>,
    pub policy_names_to_skip: Option<Vec<String>>,
    pub principal: Option<String>,
}

Fields
auth_infos: Vec<AuthInfo>

A list of authorization info objects. Simulating authorization will create a response for each authInfo object in the list.

client_id: Option<String>

The MQTT client ID.

cognito_identity_pool_id: Option<String>

The Cognito identity pool ID.

policy_names_to_add: Option<Vec<String>>

When testing custom authorization, the policies specified here are treated as if they are attached to the principal being authorized.

policy_names_to_skip: Option<Vec<String>>

When testing custom authorization, the policies specified here are treated as if they are not attached to the principal being authorized.

principal: Option<String>

The principal.