Struct rusoto_iot::TestInvokeAuthorizerResponse

pub struct TestInvokeAuthorizerResponse {
    pub disconnect_after_in_seconds: Option<i64>,
    pub is_authenticated: Option<bool>,
    pub policy_documents: Option<Vec<String>>,
    pub principal_id: Option<String>,
    pub refresh_after_in_seconds: Option<i64>,
}

Fields
disconnect_after_in_seconds: Option<i64>

The number of seconds after which the connection is terminated.

is_authenticated: Option<bool>

True if the token is authenticated, otherwise false.

policy_documents: Option<Vec<String>>

IAM policy documents.

principal_id: Option<String>

The principal ID.

refresh_after_in_seconds: Option<i64>

The number of seconds after which the temporary credentials are refreshed.