Trait rusoto_iot::Iot

pub trait Iot {
    fn accept_certificate_transfer(
        &self, 
        input: AcceptCertificateTransferRequest
    ) -> RusotoFuture<(), AcceptCertificateTransferError>;

    fn add_thing_to_thing_group(
        &self, 
        input: AddThingToThingGroupRequest
    ) -> RusotoFuture<AddThingToThingGroupResponse, AddThingToThingGroupError>;

    fn associate_targets_with_job(
        &self, 
        input: AssociateTargetsWithJobRequest
    ) -> RusotoFuture<AssociateTargetsWithJobResponse, AssociateTargetsWithJobError>;

    fn attach_policy(
        &self, 
        input: AttachPolicyRequest
    ) -> RusotoFuture<(), AttachPolicyError>;

    fn attach_principal_policy(
        &self, 
        input: AttachPrincipalPolicyRequest
    ) -> RusotoFuture<(), AttachPrincipalPolicyError>;

    fn attach_thing_principal(
        &self, 
        input: AttachThingPrincipalRequest
    ) -> RusotoFuture<AttachThingPrincipalResponse, AttachThingPrincipalError>;

    fn cancel_certificate_transfer(
        &self, 
        input: CancelCertificateTransferRequest
    ) -> RusotoFuture<(), CancelCertificateTransferError>;

    fn cancel_job(
        &self, 
        input: CancelJobRequest
    ) -> RusotoFuture<CancelJobResponse, CancelJobError>;

    fn clear_default_authorizer(
        &self
    ) -> RusotoFuture<ClearDefaultAuthorizerResponse, ClearDefaultAuthorizerError>;

    fn create_authorizer(
        &self, 
        input: CreateAuthorizerRequest
    ) -> RusotoFuture<CreateAuthorizerResponse, CreateAuthorizerError>;

    fn create_certificate_from_csr(
        &self, 
        input: CreateCertificateFromCsrRequest
    ) -> RusotoFuture<CreateCertificateFromCsrResponse, CreateCertificateFromCsrError>;

    fn create_job(
        &self, 
        input: CreateJobRequest
    ) -> RusotoFuture<CreateJobResponse, CreateJobError>;

    fn create_keys_and_certificate(
        &self, 
        input: CreateKeysAndCertificateRequest
    ) -> RusotoFuture<CreateKeysAndCertificateResponse, CreateKeysAndCertificateError>;

    fn create_ota_update(
        &self, 
        input: CreateOTAUpdateRequest
    ) -> RusotoFuture<CreateOTAUpdateResponse, CreateOTAUpdateError>;

    fn create_policy(
        &self, 
        input: CreatePolicyRequest
    ) -> RusotoFuture<CreatePolicyResponse, CreatePolicyError>;

    fn create_policy_version(
        &self, 
        input: CreatePolicyVersionRequest
    ) -> RusotoFuture<CreatePolicyVersionResponse, CreatePolicyVersionError>;

    fn create_role_alias(
        &self, 
        input: CreateRoleAliasRequest
    ) -> RusotoFuture<CreateRoleAliasResponse, CreateRoleAliasError>;

    fn create_stream(
        &self, 
        input: CreateStreamRequest
    ) -> RusotoFuture<CreateStreamResponse, CreateStreamError>;

    fn create_thing(
        &self, 
        input: CreateThingRequest
    ) -> RusotoFuture<CreateThingResponse, CreateThingError>;

    fn create_thing_group(
        &self, 
        input: CreateThingGroupRequest
    ) -> RusotoFuture<CreateThingGroupResponse, CreateThingGroupError>;

    fn create_thing_type(
        &self, 
        input: CreateThingTypeRequest
    ) -> RusotoFuture<CreateThingTypeResponse, CreateThingTypeError>;

    fn create_topic_rule(
        &self, 
        input: CreateTopicRuleRequest
    ) -> RusotoFuture<(), CreateTopicRuleError>;

    fn delete_authorizer(
        &self, 
        input: DeleteAuthorizerRequest
    ) -> RusotoFuture<DeleteAuthorizerResponse, DeleteAuthorizerError>;

    fn delete_ca_certificate(
        &self, 
        input: DeleteCACertificateRequest
    ) -> RusotoFuture<DeleteCACertificateResponse, DeleteCACertificateError>;

    fn delete_certificate(
        &self, 
        input: DeleteCertificateRequest
    ) -> RusotoFuture<(), DeleteCertificateError>;

    fn delete_ota_update(
        &self, 
        input: DeleteOTAUpdateRequest
    ) -> RusotoFuture<DeleteOTAUpdateResponse, DeleteOTAUpdateError>;

    fn delete_policy(
        &self, 
        input: DeletePolicyRequest
    ) -> RusotoFuture<(), DeletePolicyError>;

    fn delete_policy_version(
        &self, 
        input: DeletePolicyVersionRequest
    ) -> RusotoFuture<(), DeletePolicyVersionError>;

    fn delete_registration_code(
        &self
    ) -> RusotoFuture<DeleteRegistrationCodeResponse, DeleteRegistrationCodeError>;

    fn delete_role_alias(
        &self, 
        input: DeleteRoleAliasRequest
    ) -> RusotoFuture<DeleteRoleAliasResponse, DeleteRoleAliasError>;

    fn delete_stream(
        &self, 
        input: DeleteStreamRequest
    ) -> RusotoFuture<DeleteStreamResponse, DeleteStreamError>;

    fn delete_thing(
        &self, 
        input: DeleteThingRequest
    ) -> RusotoFuture<DeleteThingResponse, DeleteThingError>;

    fn delete_thing_group(
        &self, 
        input: DeleteThingGroupRequest
    ) -> RusotoFuture<DeleteThingGroupResponse, DeleteThingGroupError>;

    fn delete_thing_type(
        &self, 
        input: DeleteThingTypeRequest
    ) -> RusotoFuture<DeleteThingTypeResponse, DeleteThingTypeError>;

    fn delete_topic_rule(
        &self, 
        input: DeleteTopicRuleRequest
    ) -> RusotoFuture<(), DeleteTopicRuleError>;

    fn delete_v2_logging_level(
        &self, 
        input: DeleteV2LoggingLevelRequest
    ) -> RusotoFuture<(), DeleteV2LoggingLevelError>;

    fn deprecate_thing_type(
        &self, 
        input: DeprecateThingTypeRequest
    ) -> RusotoFuture<DeprecateThingTypeResponse, DeprecateThingTypeError>;

    fn describe_authorizer(
        &self, 
        input: DescribeAuthorizerRequest
    ) -> RusotoFuture<DescribeAuthorizerResponse, DescribeAuthorizerError>;

    fn describe_ca_certificate(
        &self, 
        input: DescribeCACertificateRequest
    ) -> RusotoFuture<DescribeCACertificateResponse, DescribeCACertificateError>;

    fn describe_certificate(
        &self, 
        input: DescribeCertificateRequest
    ) -> RusotoFuture<DescribeCertificateResponse, DescribeCertificateError>;

    fn describe_default_authorizer(
        &self
    ) -> RusotoFuture<DescribeDefaultAuthorizerResponse, DescribeDefaultAuthorizerError>;

    fn describe_endpoint(
        &self, 
        input: DescribeEndpointRequest
    ) -> RusotoFuture<DescribeEndpointResponse, DescribeEndpointError>;

    fn describe_event_configurations(
        &self
    ) -> RusotoFuture<DescribeEventConfigurationsResponse, DescribeEventConfigurationsError>;

    fn describe_index(
        &self, 
        input: DescribeIndexRequest
    ) -> RusotoFuture<DescribeIndexResponse, DescribeIndexError>;

    fn describe_job(
        &self, 
        input: DescribeJobRequest
    ) -> RusotoFuture<DescribeJobResponse, DescribeJobError>;

    fn describe_job_execution(
        &self, 
        input: DescribeJobExecutionRequest
    ) -> RusotoFuture<DescribeJobExecutionResponse, DescribeJobExecutionError>;

    fn describe_role_alias(
        &self, 
        input: DescribeRoleAliasRequest
    ) -> RusotoFuture<DescribeRoleAliasResponse, DescribeRoleAliasError>;

    fn describe_stream(
        &self, 
        input: DescribeStreamRequest
    ) -> RusotoFuture<DescribeStreamResponse, DescribeStreamError>;

    fn describe_thing(
        &self, 
        input: DescribeThingRequest
    ) -> RusotoFuture<DescribeThingResponse, DescribeThingError>;

    fn describe_thing_group(
        &self, 
        input: DescribeThingGroupRequest
    ) -> RusotoFuture<DescribeThingGroupResponse, DescribeThingGroupError>;

    fn describe_thing_registration_task(
        &self, 
        input: DescribeThingRegistrationTaskRequest
    ) -> RusotoFuture<DescribeThingRegistrationTaskResponse, DescribeThingRegistrationTaskError>;

    fn describe_thing_type(
        &self, 
        input: DescribeThingTypeRequest
    ) -> RusotoFuture<DescribeThingTypeResponse, DescribeThingTypeError>;

    fn detach_policy(
        &self, 
        input: DetachPolicyRequest
    ) -> RusotoFuture<(), DetachPolicyError>;

    fn detach_principal_policy(
        &self, 
        input: DetachPrincipalPolicyRequest
    ) -> RusotoFuture<(), DetachPrincipalPolicyError>;

    fn detach_thing_principal(
        &self, 
        input: DetachThingPrincipalRequest
    ) -> RusotoFuture<DetachThingPrincipalResponse, DetachThingPrincipalError>;

    fn disable_topic_rule(
        &self, 
        input: DisableTopicRuleRequest
    ) -> RusotoFuture<(), DisableTopicRuleError>;

    fn enable_topic_rule(
        &self, 
        input: EnableTopicRuleRequest
    ) -> RusotoFuture<(), EnableTopicRuleError>;

    fn get_effective_policies(
        &self, 
        input: GetEffectivePoliciesRequest
    ) -> RusotoFuture<GetEffectivePoliciesResponse, GetEffectivePoliciesError>;

    fn get_indexing_configuration(
        &self
    ) -> RusotoFuture<GetIndexingConfigurationResponse, GetIndexingConfigurationError>;

    fn get_job_document(
        &self, 
        input: GetJobDocumentRequest
    ) -> RusotoFuture<GetJobDocumentResponse, GetJobDocumentError>;

    fn get_logging_options(
        &self
    ) -> RusotoFuture<GetLoggingOptionsResponse, GetLoggingOptionsError>;

    fn get_ota_update(
        &self, 
        input: GetOTAUpdateRequest
    ) -> RusotoFuture<GetOTAUpdateResponse, GetOTAUpdateError>;

    fn get_policy(
        &self, 
        input: GetPolicyRequest
    ) -> RusotoFuture<GetPolicyResponse, GetPolicyError>;

    fn get_policy_version(
        &self, 
        input: GetPolicyVersionRequest
    ) -> RusotoFuture<GetPolicyVersionResponse, GetPolicyVersionError>;

    fn get_registration_code(
        &self
    ) -> RusotoFuture<GetRegistrationCodeResponse, GetRegistrationCodeError>;

    fn get_topic_rule(
        &self, 
        input: GetTopicRuleRequest
    ) -> RusotoFuture<GetTopicRuleResponse, GetTopicRuleError>;

    fn get_v2_logging_options(
        &self
    ) -> RusotoFuture<GetV2LoggingOptionsResponse, GetV2LoggingOptionsError>;

    fn list_attached_policies(
        &self, 
        input: ListAttachedPoliciesRequest
    ) -> RusotoFuture<ListAttachedPoliciesResponse, ListAttachedPoliciesError>;

    fn list_authorizers(
        &self, 
        input: ListAuthorizersRequest
    ) -> RusotoFuture<ListAuthorizersResponse, ListAuthorizersError>;

    fn list_ca_certificates(
        &self, 
        input: ListCACertificatesRequest
    ) -> RusotoFuture<ListCACertificatesResponse, ListCACertificatesError>;

    fn list_certificates(
        &self, 
        input: ListCertificatesRequest
    ) -> RusotoFuture<ListCertificatesResponse, ListCertificatesError>;

    fn list_certificates_by_ca(
        &self, 
        input: ListCertificatesByCARequest
    ) -> RusotoFuture<ListCertificatesByCAResponse, ListCertificatesByCAError>;

    fn list_indices(
        &self, 
        input: ListIndicesRequest
    ) -> RusotoFuture<ListIndicesResponse, ListIndicesError>;

    fn list_job_executions_for_job(
        &self, 
        input: ListJobExecutionsForJobRequest
    ) -> RusotoFuture<ListJobExecutionsForJobResponse, ListJobExecutionsForJobError>;

    fn list_job_executions_for_thing(
        &self, 
        input: ListJobExecutionsForThingRequest
    ) -> RusotoFuture<ListJobExecutionsForThingResponse, ListJobExecutionsForThingError>;

    fn list_jobs(
        &self, 
        input: ListJobsRequest
    ) -> RusotoFuture<ListJobsResponse, ListJobsError>;

    fn list_ota_updates(
        &self, 
        input: ListOTAUpdatesRequest
    ) -> RusotoFuture<ListOTAUpdatesResponse, ListOTAUpdatesError>;

    fn list_outgoing_certificates(
        &self, 
        input: ListOutgoingCertificatesRequest
    ) -> RusotoFuture<ListOutgoingCertificatesResponse, ListOutgoingCertificatesError>;

    fn list_policies(
        &self, 
        input: ListPoliciesRequest
    ) -> RusotoFuture<ListPoliciesResponse, ListPoliciesError>;

    fn list_policy_principals(
        &self, 
        input: ListPolicyPrincipalsRequest
    ) -> RusotoFuture<ListPolicyPrincipalsResponse, ListPolicyPrincipalsError>;

    fn list_policy_versions(
        &self, 
        input: ListPolicyVersionsRequest
    ) -> RusotoFuture<ListPolicyVersionsResponse, ListPolicyVersionsError>;

    fn list_principal_policies(
        &self, 
        input: ListPrincipalPoliciesRequest
    ) -> RusotoFuture<ListPrincipalPoliciesResponse, ListPrincipalPoliciesError>;

    fn list_principal_things(
        &self, 
        input: ListPrincipalThingsRequest
    ) -> RusotoFuture<ListPrincipalThingsResponse, ListPrincipalThingsError>;

    fn list_role_aliases(
        &self, 
        input: ListRoleAliasesRequest
    ) -> RusotoFuture<ListRoleAliasesResponse, ListRoleAliasesError>;

    fn list_streams(
        &self, 
        input: ListStreamsRequest
    ) -> RusotoFuture<ListStreamsResponse, ListStreamsError>;

    fn list_targets_for_policy(
        &self, 
        input: ListTargetsForPolicyRequest
    ) -> RusotoFuture<ListTargetsForPolicyResponse, ListTargetsForPolicyError>;

    fn list_thing_groups(
        &self, 
        input: ListThingGroupsRequest
    ) -> RusotoFuture<ListThingGroupsResponse, ListThingGroupsError>;

    fn list_thing_groups_for_thing(
        &self, 
        input: ListThingGroupsForThingRequest
    ) -> RusotoFuture<ListThingGroupsForThingResponse, ListThingGroupsForThingError>;

    fn list_thing_principals(
        &self, 
        input: ListThingPrincipalsRequest
    ) -> RusotoFuture<ListThingPrincipalsResponse, ListThingPrincipalsError>;

    fn list_thing_registration_task_reports(
        &self, 
        input: ListThingRegistrationTaskReportsRequest
    ) -> RusotoFuture<ListThingRegistrationTaskReportsResponse, ListThingRegistrationTaskReportsError>;

    fn list_thing_registration_tasks(
        &self, 
        input: ListThingRegistrationTasksRequest
    ) -> RusotoFuture<ListThingRegistrationTasksResponse, ListThingRegistrationTasksError>;

    fn list_thing_types(
        &self, 
        input: ListThingTypesRequest
    ) -> RusotoFuture<ListThingTypesResponse, ListThingTypesError>;

    fn list_things(
        &self, 
        input: ListThingsRequest
    ) -> RusotoFuture<ListThingsResponse, ListThingsError>;

    fn list_things_in_thing_group(
        &self, 
        input: ListThingsInThingGroupRequest
    ) -> RusotoFuture<ListThingsInThingGroupResponse, ListThingsInThingGroupError>;

    fn list_topic_rules(
        &self, 
        input: ListTopicRulesRequest
    ) -> RusotoFuture<ListTopicRulesResponse, ListTopicRulesError>;

    fn list_v2_logging_levels(
        &self, 
        input: ListV2LoggingLevelsRequest
    ) -> RusotoFuture<ListV2LoggingLevelsResponse, ListV2LoggingLevelsError>;

    fn register_ca_certificate(
        &self, 
        input: RegisterCACertificateRequest
    ) -> RusotoFuture<RegisterCACertificateResponse, RegisterCACertificateError>;

    fn register_certificate(
        &self, 
        input: RegisterCertificateRequest
    ) -> RusotoFuture<RegisterCertificateResponse, RegisterCertificateError>;

    fn register_thing(
        &self, 
        input: RegisterThingRequest
    ) -> RusotoFuture<RegisterThingResponse, RegisterThingError>;

    fn reject_certificate_transfer(
        &self, 
        input: RejectCertificateTransferRequest
    ) -> RusotoFuture<(), RejectCertificateTransferError>;

    fn remove_thing_from_thing_group(
        &self, 
        input: RemoveThingFromThingGroupRequest
    ) -> RusotoFuture<RemoveThingFromThingGroupResponse, RemoveThingFromThingGroupError>;

    fn replace_topic_rule(
        &self, 
        input: ReplaceTopicRuleRequest
    ) -> RusotoFuture<(), ReplaceTopicRuleError>;

    fn search_index(
        &self, 
        input: SearchIndexRequest
    ) -> RusotoFuture<SearchIndexResponse, SearchIndexError>;

    fn set_default_authorizer(
        &self, 
        input: SetDefaultAuthorizerRequest
    ) -> RusotoFuture<SetDefaultAuthorizerResponse, SetDefaultAuthorizerError>;

    fn set_default_policy_version(
        &self, 
        input: SetDefaultPolicyVersionRequest
    ) -> RusotoFuture<(), SetDefaultPolicyVersionError>;

    fn set_logging_options(
        &self, 
        input: SetLoggingOptionsRequest
    ) -> RusotoFuture<(), SetLoggingOptionsError>;

    fn set_v2_logging_level(
        &self, 
        input: SetV2LoggingLevelRequest
    ) -> RusotoFuture<(), SetV2LoggingLevelError>;

    fn set_v2_logging_options(
        &self, 
        input: SetV2LoggingOptionsRequest
    ) -> RusotoFuture<(), SetV2LoggingOptionsError>;

    fn start_thing_registration_task(
        &self, 
        input: StartThingRegistrationTaskRequest
    ) -> RusotoFuture<StartThingRegistrationTaskResponse, StartThingRegistrationTaskError>;

    fn stop_thing_registration_task(
        &self, 
        input: StopThingRegistrationTaskRequest
    ) -> RusotoFuture<StopThingRegistrationTaskResponse, StopThingRegistrationTaskError>;

    fn test_authorization(
        &self, 
        input: TestAuthorizationRequest
    ) -> RusotoFuture<TestAuthorizationResponse, TestAuthorizationError>;

    fn test_invoke_authorizer(
        &self, 
        input: TestInvokeAuthorizerRequest
    ) -> RusotoFuture<TestInvokeAuthorizerResponse, TestInvokeAuthorizerError>;

    fn transfer_certificate(
        &self, 
        input: TransferCertificateRequest
    ) -> RusotoFuture<TransferCertificateResponse, TransferCertificateError>;

    fn update_authorizer(
        &self, 
        input: UpdateAuthorizerRequest
    ) -> RusotoFuture<UpdateAuthorizerResponse, UpdateAuthorizerError>;

    fn update_ca_certificate(
        &self, 
        input: UpdateCACertificateRequest
    ) -> RusotoFuture<(), UpdateCACertificateError>;

    fn update_certificate(
        &self, 
        input: UpdateCertificateRequest
    ) -> RusotoFuture<(), UpdateCertificateError>;

    fn update_event_configurations(
        &self, 
        input: UpdateEventConfigurationsRequest
    ) -> RusotoFuture<UpdateEventConfigurationsResponse, UpdateEventConfigurationsError>;

    fn update_indexing_configuration(
        &self, 
        input: UpdateIndexingConfigurationRequest
    ) -> RusotoFuture<UpdateIndexingConfigurationResponse, UpdateIndexingConfigurationError>;

    fn update_role_alias(
        &self, 
        input: UpdateRoleAliasRequest
    ) -> RusotoFuture<UpdateRoleAliasResponse, UpdateRoleAliasError>;

    fn update_stream(
        &self, 
        input: UpdateStreamRequest
    ) -> RusotoFuture<UpdateStreamResponse, UpdateStreamError>;

    fn update_thing(
        &self, 
        input: UpdateThingRequest
    ) -> RusotoFuture<UpdateThingResponse, UpdateThingError>;

    fn update_thing_group(
        &self, 
        input: UpdateThingGroupRequest
    ) -> RusotoFuture<UpdateThingGroupResponse, UpdateThingGroupError>;

    fn update_thing_groups_for_thing(
        &self, 
        input: UpdateThingGroupsForThingRequest
    ) -> RusotoFuture<UpdateThingGroupsForThingResponse, UpdateThingGroupsForThingError>;
}

Trait representing the capabilities of the AWS IoT API. AWS IoT clients implement this trait.

Required Methods

fn accept_certificate_transfer(
&self,
input: AcceptCertificateTransferRequest
) -> RusotoFuture<(), AcceptCertificateTransferError>

Accepts a pending certificate transfer. The default state of the certificate is INACTIVE.

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

fn add_thing_to_thing_group(
&self,
input: AddThingToThingGroupRequest
) -> RusotoFuture<AddThingToThingGroupResponse, AddThingToThingGroupError>

Adds a thing to a thing group.

fn associate_targets_with_job(
&self,
input: AssociateTargetsWithJobRequest
) -> RusotoFuture<AssociateTargetsWithJobResponse, AssociateTargetsWithJobError>

Associates a group with a continuous job. The following criteria must be met:

The job must have been created with the targetSelection field set to "CONTINUOUS".

The job status must currently be "IN_PROGRESS".

The total number of targets associated with a job must not exceed 100.

fn attach_policy(
&self,
input: AttachPolicyRequest
) -> RusotoFuture<(), AttachPolicyError>

Attaches a policy to the specified target.

fn attach_principal_policy(
&self,
input: AttachPrincipalPolicyRequest
) -> RusotoFuture<(), AttachPrincipalPolicyError>

Attaches the specified policy to the specified principal (certificate or other credential).

Note: This API is deprecated. Please use AttachPolicy instead.

fn attach_thing_principal(
&self,
input: AttachThingPrincipalRequest
) -> RusotoFuture<AttachThingPrincipalResponse, AttachThingPrincipalError>

Attaches the specified principal to the specified thing.

fn cancel_certificate_transfer(
&self,
input: CancelCertificateTransferRequest
) -> RusotoFuture<(), CancelCertificateTransferError>

Cancels a pending transfer for the specified certificate.

Note Only the transfer source account can use this operation to cancel a transfer. (Transfer destinations can use RejectCertificateTransfer instead.) After transfer, AWS IoT returns the certificate to the source account in the INACTIVE state. After the destination account has accepted the transfer, the transfer cannot be cancelled.

After a certificate transfer is cancelled, the status of the certificate changes from PENDING_TRANSFER to INACTIVE.

fn cancel_job(
&self,
input: CancelJobRequest
) -> RusotoFuture<CancelJobResponse, CancelJobError>

Cancels a job.

fn clear_default_authorizer(
&self
) -> RusotoFuture<ClearDefaultAuthorizerResponse, ClearDefaultAuthorizerError>

Clears the default authorizer.

fn create_authorizer(
&self,
input: CreateAuthorizerRequest
) -> RusotoFuture<CreateAuthorizerResponse, CreateAuthorizerError>

Creates an authorizer.

fn create_certificate_from_csr(
&self,
input: CreateCertificateFromCsrRequest
) -> RusotoFuture<CreateCertificateFromCsrResponse, CreateCertificateFromCsrError>

Creates an X.509 certificate using the specified certificate signing request.

Note: The CSR must include a public key that is either an RSA key with a length of at least 2048 bits or an ECC key from NIST P-256 or NIST P-384 curves.

Note: Reusing the same certificate signing request (CSR) results in a distinct certificate.

You can create multiple certificates in a batch by creating a directory, copying multiple .csr files into that directory, and then specifying that directory on the command line. The following commands show how to create a batch of certificates given a batch of CSRs.

Assuming a set of CSRs are located inside of the directory my-csr-directory:

On Linux and OS X, the command is:

$ ls my-csr-directory/ | xargs -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

This command lists all of the CSRs in my-csr-directory and pipes each CSR file name to the aws iot create-certificate-from-csr AWS CLI command to create a certificate for the corresponding CSR.

The aws iot create-certificate-from-csr part of the command can also be run in parallel to speed up the certificate creation process:

$ ls my-csr-directory/ | xargs -P 10 -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

On Windows PowerShell, the command to create certificates for all CSRs in my-csr-directory is:

> ls -Name my-csr-directory | %{aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/$_}

On a Windows command prompt, the command to create certificates for all CSRs in my-csr-directory is:

> forfiles /p my-csr-directory /c "cmd /c aws iot create-certificate-from-csr --certificate-signing-request file://@path"

fn create_job(
&self,
input: CreateJobRequest
) -> RusotoFuture<CreateJobResponse, CreateJobError>

Creates a job.

fn create_keys_and_certificate(
&self,
input: CreateKeysAndCertificateRequest
) -> RusotoFuture<CreateKeysAndCertificateResponse, CreateKeysAndCertificateError>

Creates a 2048-bit RSA key pair and issues an X.509 certificate using the issued public key.

Note This is the only time AWS IoT issues the private key for this certificate, so it is important to keep it in a secure location.

fn create_ota_update(
&self,
input: CreateOTAUpdateRequest
) -> RusotoFuture<CreateOTAUpdateResponse, CreateOTAUpdateError>

Creates an AWS IoT OTAUpdate on a target group of things or groups.

fn create_policy(
&self,
input: CreatePolicyRequest
) -> RusotoFuture<CreatePolicyResponse, CreatePolicyError>

Creates an AWS IoT policy.

The created policy is the default version for the policy. This operation creates a policy version with a version identifier of 1 and sets 1 as the policy's default version.

fn create_policy_version(
&self,
input: CreatePolicyVersionRequest
) -> RusotoFuture<CreatePolicyVersionResponse, CreatePolicyVersionError>

Creates a new version of the specified AWS IoT policy. To update a policy, create a new policy version. A managed policy can have up to five versions. If the policy has five versions, you must use DeletePolicyVersion to delete an existing version before you create a new one.

Optionally, you can set the new version as the policy's default version. The default version is the operative version (that is, the version that is in effect for the certificates to which the policy is attached).

fn create_role_alias(
&self,
input: CreateRoleAliasRequest
) -> RusotoFuture<CreateRoleAliasResponse, CreateRoleAliasError>

Creates a role alias.

fn create_stream(
&self,
input: CreateStreamRequest
) -> RusotoFuture<CreateStreamResponse, CreateStreamError>

Creates a stream for delivering one or more large files in chunks over MQTT. A stream transports data bytes in chunks or blocks packaged as MQTT messages from a source like S3. You can have one or more files associated with a stream. The total size of a file associated with the stream cannot exceed more than 2 MB. The stream will be created with version 0. If a stream is created with the same streamID as a stream that existed and was deleted within last 90 days, we will resurrect that old stream by incrementing the version by 1.

fn create_thing(
&self,
input: CreateThingRequest
) -> RusotoFuture<CreateThingResponse, CreateThingError>

Creates a thing record in the thing registry.

fn create_thing_group(
&self,
input: CreateThingGroupRequest
) -> RusotoFuture<CreateThingGroupResponse, CreateThingGroupError>

Create a thing group.

fn create_thing_type(
&self,
input: CreateThingTypeRequest
) -> RusotoFuture<CreateThingTypeResponse, CreateThingTypeError>

Creates a new thing type.

fn create_topic_rule(
&self,
input: CreateTopicRuleRequest
) -> RusotoFuture<(), CreateTopicRuleError>

Creates a rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.

fn delete_authorizer(
&self,
input: DeleteAuthorizerRequest
) -> RusotoFuture<DeleteAuthorizerResponse, DeleteAuthorizerError>

Deletes an authorizer.

fn delete_ca_certificate(
&self,
input: DeleteCACertificateRequest
) -> RusotoFuture<DeleteCACertificateResponse, DeleteCACertificateError>

Deletes a registered CA certificate.

fn delete_certificate(
&self,
input: DeleteCertificateRequest
) -> RusotoFuture<(), DeleteCertificateError>

Deletes the specified certificate.

A certificate cannot be deleted if it has a policy attached to it or if its status is set to ACTIVE. To delete a certificate, first use the DetachPrincipalPolicy API to detach all policies. Next, use the UpdateCertificate API to set the certificate to the INACTIVE status.

fn delete_ota_update(
&self,
input: DeleteOTAUpdateRequest
) -> RusotoFuture<DeleteOTAUpdateResponse, DeleteOTAUpdateError>

Delete an OTA update.

fn delete_policy(
&self,
input: DeletePolicyRequest
) -> RusotoFuture<(), DeletePolicyError>

Deletes the specified policy.

A policy cannot be deleted if it has non-default versions or it is attached to any certificate.

To delete a policy, use the DeletePolicyVersion API to delete all non-default versions of the policy; use the DetachPrincipalPolicy API to detach the policy from any certificate; and then use the DeletePolicy API to delete the policy.

When a policy is deleted using DeletePolicy, its default version is deleted with it.

fn delete_policy_version(
&self,
input: DeletePolicyVersionRequest
) -> RusotoFuture<(), DeletePolicyVersionError>

Deletes the specified version of the specified policy. You cannot delete the default version of a policy using this API. To delete the default version of a policy, use DeletePolicy. To find out which version of a policy is marked as the default version, use ListPolicyVersions.

fn delete_registration_code(
&self
) -> RusotoFuture<DeleteRegistrationCodeResponse, DeleteRegistrationCodeError>

Deletes a CA certificate registration code.

fn delete_role_alias(
&self,
input: DeleteRoleAliasRequest
) -> RusotoFuture<DeleteRoleAliasResponse, DeleteRoleAliasError>

Deletes a role alias

fn delete_stream(
&self,
input: DeleteStreamRequest
) -> RusotoFuture<DeleteStreamResponse, DeleteStreamError>

Deletes a stream.

fn delete_thing(
&self,
input: DeleteThingRequest
) -> RusotoFuture<DeleteThingResponse, DeleteThingError>

Deletes the specified thing.

fn delete_thing_group(
&self,
input: DeleteThingGroupRequest
) -> RusotoFuture<DeleteThingGroupResponse, DeleteThingGroupError>

Deletes a thing group.

fn delete_thing_type(
&self,
input: DeleteThingTypeRequest
) -> RusotoFuture<DeleteThingTypeResponse, DeleteThingTypeError>

Deletes the specified thing type . You cannot delete a thing type if it has things associated with it. To delete a thing type, first mark it as deprecated by calling DeprecateThingType, then remove any associated things by calling UpdateThing to change the thing type on any associated thing, and finally use DeleteThingType to delete the thing type.

fn delete_topic_rule(
&self,
input: DeleteTopicRuleRequest
) -> RusotoFuture<(), DeleteTopicRuleError>

Deletes the rule.

fn delete_v2_logging_level(
&self,
input: DeleteV2LoggingLevelRequest
) -> RusotoFuture<(), DeleteV2LoggingLevelError>

Deletes a logging level.

fn deprecate_thing_type(
&self,
input: DeprecateThingTypeRequest
) -> RusotoFuture<DeprecateThingTypeResponse, DeprecateThingTypeError>

Deprecates a thing type. You can not associate new things with deprecated thing type.

fn describe_authorizer(
&self,
input: DescribeAuthorizerRequest
) -> RusotoFuture<DescribeAuthorizerResponse, DescribeAuthorizerError>

Describes an authorizer.

fn describe_ca_certificate(
&self,
input: DescribeCACertificateRequest
) -> RusotoFuture<DescribeCACertificateResponse, DescribeCACertificateError>

Describes a registered CA certificate.

fn describe_certificate(
&self,
input: DescribeCertificateRequest
) -> RusotoFuture<DescribeCertificateResponse, DescribeCertificateError>

Gets information about the specified certificate.

fn describe_default_authorizer(
&self
) -> RusotoFuture<DescribeDefaultAuthorizerResponse, DescribeDefaultAuthorizerError>

Describes the default authorizer.

fn describe_endpoint(
&self,
input: DescribeEndpointRequest
) -> RusotoFuture<DescribeEndpointResponse, DescribeEndpointError>

Returns a unique endpoint specific to the AWS account making the call.

fn describe_event_configurations(
&self
) -> RusotoFuture<DescribeEventConfigurationsResponse, DescribeEventConfigurationsError>

Describes event configurations.

fn describe_index(
&self,
input: DescribeIndexRequest
) -> RusotoFuture<DescribeIndexResponse, DescribeIndexError>

Describes a search index.

fn describe_job(
&self,
input: DescribeJobRequest
) -> RusotoFuture<DescribeJobResponse, DescribeJobError>

Describes a job.

fn describe_job_execution(
&self,
input: DescribeJobExecutionRequest
) -> RusotoFuture<DescribeJobExecutionResponse, DescribeJobExecutionError>

Describes a job execution.

fn describe_role_alias(
&self,
input: DescribeRoleAliasRequest
) -> RusotoFuture<DescribeRoleAliasResponse, DescribeRoleAliasError>

Describes a role alias.

fn describe_stream(
&self,
input: DescribeStreamRequest
) -> RusotoFuture<DescribeStreamResponse, DescribeStreamError>

Gets information about a stream.

fn describe_thing(
&self,
input: DescribeThingRequest
) -> RusotoFuture<DescribeThingResponse, DescribeThingError>

Gets information about the specified thing.

fn describe_thing_group(
&self,
input: DescribeThingGroupRequest
) -> RusotoFuture<DescribeThingGroupResponse, DescribeThingGroupError>

Describe a thing group.

fn describe_thing_registration_task(
&self,
input: DescribeThingRegistrationTaskRequest
) -> RusotoFuture<DescribeThingRegistrationTaskResponse, DescribeThingRegistrationTaskError>

Describes a bulk thing provisioning task.

fn describe_thing_type(
&self,
input: DescribeThingTypeRequest
) -> RusotoFuture<DescribeThingTypeResponse, DescribeThingTypeError>

Gets information about the specified thing type.

fn detach_policy(
&self,
input: DetachPolicyRequest
) -> RusotoFuture<(), DetachPolicyError>

Detaches a policy from the specified target.

fn detach_principal_policy(
&self,
input: DetachPrincipalPolicyRequest
) -> RusotoFuture<(), DetachPrincipalPolicyError>

Removes the specified policy from the specified certificate.

Note: This API is deprecated. Please use DetachPolicy instead.

fn detach_thing_principal(
&self,
input: DetachThingPrincipalRequest
) -> RusotoFuture<DetachThingPrincipalResponse, DetachThingPrincipalError>

Detaches the specified principal from the specified thing.

fn disable_topic_rule(
&self,
input: DisableTopicRuleRequest
) -> RusotoFuture<(), DisableTopicRuleError>

Disables the rule.

fn enable_topic_rule(
&self,
input: EnableTopicRuleRequest
) -> RusotoFuture<(), EnableTopicRuleError>

Enables the rule.

fn get_effective_policies(
&self,
input: GetEffectivePoliciesRequest
) -> RusotoFuture<GetEffectivePoliciesResponse, GetEffectivePoliciesError>

Gets effective policies.

fn get_indexing_configuration(
&self
) -> RusotoFuture<GetIndexingConfigurationResponse, GetIndexingConfigurationError>

Gets the search configuration.

fn get_job_document(
&self,
input: GetJobDocumentRequest
) -> RusotoFuture<GetJobDocumentResponse, GetJobDocumentError>

Gets a job document.

fn get_logging_options(
&self
) -> RusotoFuture<GetLoggingOptionsResponse, GetLoggingOptionsError>

Gets the logging options.

fn get_ota_update(
&self,
input: GetOTAUpdateRequest
) -> RusotoFuture<GetOTAUpdateResponse, GetOTAUpdateError>

Gets an OTA update.

fn get_policy(
&self,
input: GetPolicyRequest
) -> RusotoFuture<GetPolicyResponse, GetPolicyError>

Gets information about the specified policy with the policy document of the default version.

fn get_policy_version(
&self,
input: GetPolicyVersionRequest
) -> RusotoFuture<GetPolicyVersionResponse, GetPolicyVersionError>

Gets information about the specified policy version.

fn get_registration_code(
&self
) -> RusotoFuture<GetRegistrationCodeResponse, GetRegistrationCodeError>

Gets a registration code used to register a CA certificate with AWS IoT.

fn get_topic_rule(
&self,
input: GetTopicRuleRequest
) -> RusotoFuture<GetTopicRuleResponse, GetTopicRuleError>

Gets information about the rule.

fn get_v2_logging_options(
&self
) -> RusotoFuture<GetV2LoggingOptionsResponse, GetV2LoggingOptionsError>

Gets the fine grained logging options.

fn list_attached_policies(
&self,
input: ListAttachedPoliciesRequest
) -> RusotoFuture<ListAttachedPoliciesResponse, ListAttachedPoliciesError>

Lists the policies attached to the specified thing group.

fn list_authorizers(
&self,
input: ListAuthorizersRequest
) -> RusotoFuture<ListAuthorizersResponse, ListAuthorizersError>

Lists the authorizers registered in your account.

fn list_ca_certificates(
&self,
input: ListCACertificatesRequest
) -> RusotoFuture<ListCACertificatesResponse, ListCACertificatesError>

Lists the CA certificates registered for your AWS account.

The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.

fn list_certificates(
&self,
input: ListCertificatesRequest
) -> RusotoFuture<ListCertificatesResponse, ListCertificatesError>

Lists the certificates registered in your AWS account.

The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.

fn list_certificates_by_ca(
&self,
input: ListCertificatesByCARequest
) -> RusotoFuture<ListCertificatesByCAResponse, ListCertificatesByCAError>

List the device certificates signed by the specified CA certificate.

fn list_indices(
&self,
input: ListIndicesRequest
) -> RusotoFuture<ListIndicesResponse, ListIndicesError>

Lists the search indices.

fn list_job_executions_for_job(
&self,
input: ListJobExecutionsForJobRequest
) -> RusotoFuture<ListJobExecutionsForJobResponse, ListJobExecutionsForJobError>

Lists the job executions for a job.

fn list_job_executions_for_thing(
&self,
input: ListJobExecutionsForThingRequest
) -> RusotoFuture<ListJobExecutionsForThingResponse, ListJobExecutionsForThingError>

Lists the job executions for the specified thing.

fn list_jobs(
&self,
input: ListJobsRequest
) -> RusotoFuture<ListJobsResponse, ListJobsError>

Lists jobs.

fn list_ota_updates(
&self,
input: ListOTAUpdatesRequest
) -> RusotoFuture<ListOTAUpdatesResponse, ListOTAUpdatesError>

Lists OTA updates.

fn list_outgoing_certificates(
&self,
input: ListOutgoingCertificatesRequest
) -> RusotoFuture<ListOutgoingCertificatesResponse, ListOutgoingCertificatesError>

Lists certificates that are being transferred but not yet accepted.

fn list_policies(
&self,
input: ListPoliciesRequest
) -> RusotoFuture<ListPoliciesResponse, ListPoliciesError>

Lists your policies.

fn list_policy_principals(
&self,
input: ListPolicyPrincipalsRequest
) -> RusotoFuture<ListPolicyPrincipalsResponse, ListPolicyPrincipalsError>

Lists the principals associated with the specified policy.

Note: This API is deprecated. Please use ListTargetsForPolicy instead.

fn list_policy_versions(
&self,
input: ListPolicyVersionsRequest
) -> RusotoFuture<ListPolicyVersionsResponse, ListPolicyVersionsError>

Lists the versions of the specified policy and identifies the default version.

fn list_principal_policies(
&self,
input: ListPrincipalPoliciesRequest
) -> RusotoFuture<ListPrincipalPoliciesResponse, ListPrincipalPoliciesError>

Lists the policies attached to the specified principal. If you use an Cognito identity, the ID must be in AmazonCognito Identity format.

Note: This API is deprecated. Please use ListAttachedPolicies instead.

fn list_principal_things(
&self,
input: ListPrincipalThingsRequest
) -> RusotoFuture<ListPrincipalThingsResponse, ListPrincipalThingsError>

Lists the things associated with the specified principal.

fn list_role_aliases(
&self,
input: ListRoleAliasesRequest
) -> RusotoFuture<ListRoleAliasesResponse, ListRoleAliasesError>

Lists the role aliases registered in your account.

fn list_streams(
&self,
input: ListStreamsRequest
) -> RusotoFuture<ListStreamsResponse, ListStreamsError>

Lists all of the streams in your AWS account.

fn list_targets_for_policy(
&self,
input: ListTargetsForPolicyRequest
) -> RusotoFuture<ListTargetsForPolicyResponse, ListTargetsForPolicyError>

List targets for the specified policy.

fn list_thing_groups(
&self,
input: ListThingGroupsRequest
) -> RusotoFuture<ListThingGroupsResponse, ListThingGroupsError>

List the thing groups in your account.

fn list_thing_groups_for_thing(
&self,
input: ListThingGroupsForThingRequest
) -> RusotoFuture<ListThingGroupsForThingResponse, ListThingGroupsForThingError>

List the thing groups to which the specified thing belongs.

fn list_thing_principals(
&self,
input: ListThingPrincipalsRequest
) -> RusotoFuture<ListThingPrincipalsResponse, ListThingPrincipalsError>

Lists the principals associated with the specified thing.

fn list_thing_registration_task_reports(
&self,
input: ListThingRegistrationTaskReportsRequest
) -> RusotoFuture<ListThingRegistrationTaskReportsResponse, ListThingRegistrationTaskReportsError>

Information about the thing registration tasks.

fn list_thing_registration_tasks(
&self,
input: ListThingRegistrationTasksRequest
) -> RusotoFuture<ListThingRegistrationTasksResponse, ListThingRegistrationTasksError>

List bulk thing provisioning tasks.

fn list_thing_types(
&self,
input: ListThingTypesRequest
) -> RusotoFuture<ListThingTypesResponse, ListThingTypesError>

Lists the existing thing types.

fn list_things(
&self,
input: ListThingsRequest
) -> RusotoFuture<ListThingsResponse, ListThingsError>

Lists your things. Use the attributeName and attributeValue parameters to filter your things. For example, calling ListThings with attributeName=Color and attributeValue=Red retrieves all things in the registry that contain an attribute Color with the value Red.

fn list_things_in_thing_group(
&self,
input: ListThingsInThingGroupRequest
) -> RusotoFuture<ListThingsInThingGroupResponse, ListThingsInThingGroupError>

Lists the things in the specified group.

fn list_topic_rules(
&self,
input: ListTopicRulesRequest
) -> RusotoFuture<ListTopicRulesResponse, ListTopicRulesError>

Lists the rules for the specific topic.

fn list_v2_logging_levels(
&self,
input: ListV2LoggingLevelsRequest
) -> RusotoFuture<ListV2LoggingLevelsResponse, ListV2LoggingLevelsError>

Lists logging levels.

fn register_ca_certificate(
&self,
input: RegisterCACertificateRequest
) -> RusotoFuture<RegisterCACertificateResponse, RegisterCACertificateError>

Registers a CA certificate with AWS IoT. This CA certificate can then be used to sign device certificates, which can be then registered with AWS IoT. You can register up to 10 CA certificates per AWS account that have the same subject field. This enables you to have up to 10 certificate authorities sign your device certificates. If you have more than one CA certificate registered, make sure you pass the CA certificate when you register your device certificates with the RegisterCertificate API.

fn register_certificate(
&self,
input: RegisterCertificateRequest
) -> RusotoFuture<RegisterCertificateResponse, RegisterCertificateError>

Registers a device certificate with AWS IoT. If you have more than one CA certificate that has the same subject field, you must specify the CA certificate that was used to sign the device certificate being registered.

fn register_thing(
&self,
input: RegisterThingRequest
) -> RusotoFuture<RegisterThingResponse, RegisterThingError>

Provisions a thing.

fn reject_certificate_transfer(
&self,
input: RejectCertificateTransferRequest
) -> RusotoFuture<(), RejectCertificateTransferError>

Rejects a pending certificate transfer. After AWS IoT rejects a certificate transfer, the certificate status changes from PENDING_TRANSFER to INACTIVE.

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

This operation can only be called by the transfer destination. After it is called, the certificate will be returned to the source's account in the INACTIVE state.

fn remove_thing_from_thing_group(
&self,
input: RemoveThingFromThingGroupRequest
) -> RusotoFuture<RemoveThingFromThingGroupResponse, RemoveThingFromThingGroupError>

Remove the specified thing from the specified group.

fn replace_topic_rule(
&self,
input: ReplaceTopicRuleRequest
) -> RusotoFuture<(), ReplaceTopicRuleError>

Replaces the rule. You must specify all parameters for the new rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.

fn search_index(
&self,
input: SearchIndexRequest
) -> RusotoFuture<SearchIndexResponse, SearchIndexError>

The query search index.

fn set_default_authorizer(
&self,
input: SetDefaultAuthorizerRequest
) -> RusotoFuture<SetDefaultAuthorizerResponse, SetDefaultAuthorizerError>

Sets the default authorizer. This will be used if a websocket connection is made without specifying an authorizer.

fn set_default_policy_version(
&self,
input: SetDefaultPolicyVersionRequest
) -> RusotoFuture<(), SetDefaultPolicyVersionError>

Sets the specified version of the specified policy as the policy's default (operative) version. This action affects all certificates to which the policy is attached. To list the principals the policy is attached to, use the ListPrincipalPolicy API.

fn set_logging_options(
&self,
input: SetLoggingOptionsRequest
) -> RusotoFuture<(), SetLoggingOptionsError>

Sets the logging options.

fn set_v2_logging_level(
&self,
input: SetV2LoggingLevelRequest
) -> RusotoFuture<(), SetV2LoggingLevelError>

Sets the logging level.

fn set_v2_logging_options(
&self,
input: SetV2LoggingOptionsRequest
) -> RusotoFuture<(), SetV2LoggingOptionsError>

Sets the logging options for the V2 logging service.

fn start_thing_registration_task(
&self,
input: StartThingRegistrationTaskRequest
) -> RusotoFuture<StartThingRegistrationTaskResponse, StartThingRegistrationTaskError>

Creates a bulk thing provisioning task.

fn stop_thing_registration_task(
&self,
input: StopThingRegistrationTaskRequest
) -> RusotoFuture<StopThingRegistrationTaskResponse, StopThingRegistrationTaskError>

Cancels a bulk thing provisioning task.

fn test_authorization(
&self,
input: TestAuthorizationRequest
) -> RusotoFuture<TestAuthorizationResponse, TestAuthorizationError>

Test custom authorization.

fn test_invoke_authorizer(
&self,
input: TestInvokeAuthorizerRequest
) -> RusotoFuture<TestInvokeAuthorizerResponse, TestInvokeAuthorizerError>

Invoke the specified custom authorizer for testing purposes.

fn transfer_certificate(
&self,
input: TransferCertificateRequest
) -> RusotoFuture<TransferCertificateResponse, TransferCertificateError>

Transfers the specified certificate to the specified AWS account.

You can cancel the transfer until it is acknowledged by the recipient.

No notification is sent to the transfer destination's account. It is up to the caller to notify the transfer target.

The certificate being transferred must not be in the ACTIVE state. You can use the UpdateCertificate API to deactivate it.

The certificate must not have any policies attached to it. You can use the DetachPrincipalPolicy API to detach them.

fn update_authorizer(
&self,
input: UpdateAuthorizerRequest
) -> RusotoFuture<UpdateAuthorizerResponse, UpdateAuthorizerError>

Updates an authorizer.

fn update_ca_certificate(
&self,
input: UpdateCACertificateRequest
) -> RusotoFuture<(), UpdateCACertificateError>

Updates a registered CA certificate.

fn update_certificate(
&self,
input: UpdateCertificateRequest
) -> RusotoFuture<(), UpdateCertificateError>

Updates the status of the specified certificate. This operation is idempotent.

Moving a certificate from the ACTIVE state (including REVOKED) will not disconnect currently connected devices, but these devices will be unable to reconnect.

The ACTIVE state is required to authenticate devices connecting to AWS IoT using a certificate.

fn update_event_configurations(
&self,
input: UpdateEventConfigurationsRequest
) -> RusotoFuture<UpdateEventConfigurationsResponse, UpdateEventConfigurationsError>

Updates the event configurations.

fn update_indexing_configuration(
&self,
input: UpdateIndexingConfigurationRequest
) -> RusotoFuture<UpdateIndexingConfigurationResponse, UpdateIndexingConfigurationError>

Updates the search configuration.

fn update_role_alias(
&self,
input: UpdateRoleAliasRequest
) -> RusotoFuture<UpdateRoleAliasResponse, UpdateRoleAliasError>

Updates a role alias.

fn update_stream(
&self,
input: UpdateStreamRequest
) -> RusotoFuture<UpdateStreamResponse, UpdateStreamError>

Updates an existing stream. The stream version will be incremented by one.

fn update_thing(
&self,
input: UpdateThingRequest
) -> RusotoFuture<UpdateThingResponse, UpdateThingError>

Updates the data for a thing.

fn update_thing_group(
&self,
input: UpdateThingGroupRequest
) -> RusotoFuture<UpdateThingGroupResponse, UpdateThingGroupError>

Update a thing group.

fn update_thing_groups_for_thing(
&self,
input: UpdateThingGroupsForThingRequest
) -> RusotoFuture<UpdateThingGroupsForThingResponse, UpdateThingGroupsForThingError>

Updates the groups to which the thing belongs.