Crate rusoto_iot

AWS IoT
AWS IoT provides secure, bi-directional communication between Internet-connected things (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud. You can discover your custom IoT-Data endpoint to communicate with, configure rules for data processing and integration with other services, organize resources associated with each thing (Thing Registry), configure logging, and create and manage policies and credentials to authenticate things.

For more information about how AWS IoT works, see the Developer Guide.

If you're using the service, you're probably looking for IotClient and Iot.

Structs

AcceptCertificateTransferRequest
The input for the AcceptCertificateTransfer operation.

Action
Describes the actions associated with a rule.

AttachPrincipalPolicyRequest
The input for the AttachPrincipalPolicy operation.

AttachThingPrincipalRequest
The input for the AttachThingPrincipal operation.

AttachThingPrincipalResponse
The output from the AttachThingPrincipal operation.

AttributePayload
The attribute payload.

CACertificate
A CA certificate.

CACertificateDescription
Describes a CA certificate.

CancelCertificateTransferRequest
The input for the CancelCertificateTransfer operation.

Certificate
Information about a certificate.

CertificateDescription
Describes a certificate.

CloudwatchAlarmAction
Describes an action that updates a CloudWatch alarm.

CloudwatchMetricAction
Describes an action that captures a CloudWatch metric.

CreateCertificateFromCsrRequest
The input for the CreateCertificateFromCsr operation.

CreateCertificateFromCsrResponse
The output from the CreateCertificateFromCsr operation.

CreateKeysAndCertificateRequest
The input for the CreateKeysAndCertificate operation.

CreateKeysAndCertificateResponse
The output of the CreateKeysAndCertificate operation.

CreatePolicyRequest
The input for the CreatePolicy operation.

CreatePolicyResponse
The output from the CreatePolicy operation.

CreatePolicyVersionRequest
The input for the CreatePolicyVersion operation.

CreatePolicyVersionResponse
The output of the CreatePolicyVersion operation.

CreateThingRequest
The input for the CreateThing operation.

CreateThingResponse
The output of the CreateThing operation.

CreateThingTypeRequest
The input for the CreateThingType operation.

CreateThingTypeResponse
The output of the CreateThingType operation.

CreateTopicRuleRequest
The input for the CreateTopicRule operation.

DeleteCACertificateRequest
Input for the DeleteCACertificate operation.

DeleteCACertificateResponse
The output for the DeleteCACertificate operation.

DeleteCertificateRequest
The input for the DeleteCertificate operation.

DeletePolicyRequest
The input for the DeletePolicy operation.

DeletePolicyVersionRequest
The input for the DeletePolicyVersion operation.

DeleteRegistrationCodeRequest
The input for the DeleteRegistrationCode operation.

DeleteRegistrationCodeResponse
The output for the DeleteRegistrationCode operation.

DeleteThingRequest
The input for the DeleteThing operation.

DeleteThingResponse
The output of the DeleteThing operation.

DeleteThingTypeRequest
The input for the DeleteThingType operation.

DeleteThingTypeResponse
The output for the DeleteThingType operation.

DeleteTopicRuleRequest
The input for the DeleteTopicRule operation.

DeprecateThingTypeRequest
The input for the DeprecateThingType operation.

DeprecateThingTypeResponse
The output for the DeprecateThingType operation.

DescribeCACertificateRequest
The input for the DescribeCACertificate operation.

DescribeCACertificateResponse
The output from the DescribeCACertificate operation.

DescribeCertificateRequest
The input for the DescribeCertificate operation.

DescribeCertificateResponse
The output of the DescribeCertificate operation.

DescribeEndpointRequest
The input for the DescribeEndpoint operation.

DescribeEndpointResponse
The output from the DescribeEndpoint operation.

DescribeThingRequest
The input for the DescribeThing operation.

DescribeThingResponse
The output from the DescribeThing operation.

DescribeThingTypeRequest
The input for the DescribeThingType operation.

DescribeThingTypeResponse
The output for the DescribeThingType operation.

DetachPrincipalPolicyRequest
The input for the DetachPrincipalPolicy operation.

DetachThingPrincipalRequest
The input for the DetachThingPrincipal operation.

DetachThingPrincipalResponse
The output from the DetachThingPrincipal operation.

DisableTopicRuleRequest
The input for the DisableTopicRuleRequest operation.

DynamoDBAction
Describes an action to write to a DynamoDB table.

The tableName, hashKeyField, and rangeKeyField values must match the values used when you created the table.

The hashKeyValue and rangeKeyvalue fields use a substitution template syntax. These templates provide data at runtime. The syntax is as follows: ${sql-expression}.

You can specify any valid expression in a WHERE or SELECT clause, including JSON properties, comparisons, calculations, and functions. For example, the following field uses the third level of the topic:

"hashKeyValue": "${topic(3)}"

The following field uses the timestamp:

"rangeKeyValue": "${timestamp()}"

DynamoDBv2Action
Describes an action to write to a DynamoDB table.

This DynamoDB action writes each attribute in the message payload into it's own column in the DynamoDB table.

ElasticsearchAction
Describes an action that writes data to an Amazon Elasticsearch Service domain.

EnableTopicRuleRequest
The input for the EnableTopicRuleRequest operation.

FirehoseAction
Describes an action that writes data to an Amazon Kinesis Firehose stream.

GetLoggingOptionsRequest
The input for the GetLoggingOptions operation.

GetLoggingOptionsResponse
The output from the GetLoggingOptions operation.

GetPolicyRequest
The input for the GetPolicy operation.

GetPolicyResponse
The output from the GetPolicy operation.

GetPolicyVersionRequest
The input for the GetPolicyVersion operation.

GetPolicyVersionResponse
The output from the GetPolicyVersion operation.

GetRegistrationCodeRequest
The input to the GetRegistrationCode operation.

GetRegistrationCodeResponse
The output from the GetRegistrationCode operation.

GetTopicRuleRequest
The input for the GetTopicRule operation.

GetTopicRuleResponse
The output from the GetTopicRule operation.

IotClient
A client for the AWS IoT API.

KeyPair
Describes a key pair.

KinesisAction
Describes an action to write data to an Amazon Kinesis stream.

LambdaAction
Describes an action to invoke a Lambda function.

ListCACertificatesRequest
Input for the ListCACertificates operation.

ListCACertificatesResponse
The output from the ListCACertificates operation.

ListCertificatesByCARequest
The input to the ListCertificatesByCA operation.

ListCertificatesByCAResponse
The output of the ListCertificatesByCA operation.

ListCertificatesRequest
The input for the ListCertificates operation.

ListCertificatesResponse
The output of the ListCertificates operation.

ListOutgoingCertificatesRequest
The input to the ListOutgoingCertificates operation.

ListOutgoingCertificatesResponse
The output from the ListOutgoingCertificates operation.

ListPoliciesRequest
The input for the ListPolicies operation.

ListPoliciesResponse
The output from the ListPolicies operation.

ListPolicyPrincipalsRequest
The input for the ListPolicyPrincipals operation.

ListPolicyPrincipalsResponse
The output from the ListPolicyPrincipals operation.

ListPolicyVersionsRequest
The input for the ListPolicyVersions operation.

ListPolicyVersionsResponse
The output from the ListPolicyVersions operation.

ListPrincipalPoliciesRequest
The input for the ListPrincipalPolicies operation.

ListPrincipalPoliciesResponse
The output from the ListPrincipalPolicies operation.

ListPrincipalThingsRequest
The input for the ListPrincipalThings operation.

ListPrincipalThingsResponse
The output from the ListPrincipalThings operation.

ListThingPrincipalsRequest
The input for the ListThingPrincipal operation.

ListThingPrincipalsResponse
The output from the ListThingPrincipals operation.

ListThingTypesRequest
The input for the ListThingTypes operation.

ListThingTypesResponse
The output for the ListThingTypes operation.

ListThingsRequest
The input for the ListThings operation.

ListThingsResponse
The output from the ListThings operation.

ListTopicRulesRequest
The input for the ListTopicRules operation.

ListTopicRulesResponse
The output from the ListTopicRules operation.

LoggingOptionsPayload
Describes the logging options payload.

OutgoingCertificate
A certificate that has been transfered but not yet accepted.

Policy
Describes an AWS IoT policy.

PolicyVersion
Describes a policy version.

PutItemInput
The input for the DynamoActionVS action that specifies the DynamoDB table to which the message data will be written.

RegisterCACertificateRequest
The input to the RegisterCACertificate operation.

RegisterCACertificateResponse
The output from the RegisterCACertificateResponse operation.

RegisterCertificateRequest
The input to the RegisterCertificate operation.

RegisterCertificateResponse
The output from the RegisterCertificate operation.

RejectCertificateTransferRequest
The input for the RejectCertificateTransfer operation.

ReplaceTopicRuleRequest
The input for the ReplaceTopicRule operation.

RepublishAction
Describes an action to republish to another topic.

S3Action
Describes an action to write data to an Amazon S3 bucket.

SalesforceAction
Describes an action to write a message to a Salesforce IoT Cloud Input Stream.

SetDefaultPolicyVersionRequest
The input for the SetDefaultPolicyVersion operation.

SetLoggingOptionsRequest
The input for the SetLoggingOptions operation.

SnsAction
Describes an action to publish to an Amazon SNS topic.

SqsAction
Describes an action to publish data to an Amazon SQS queue.

ThingAttribute
The properties of the thing, including thing name, thing type name, and a list of thing attributes.

ThingTypeDefinition
The definition of the thing type, including thing type name and description.

ThingTypeMetadata
The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when time was deprecated.

ThingTypeProperties
The ThingTypeProperties contains information about the thing type including: a thing type description, and a list of searchable thing attribute names.

TopicRule
Describes a rule.

TopicRuleListItem
Describes a rule.

TopicRulePayload
Describes a rule.

TransferCertificateRequest
The input for the TransferCertificate operation.

TransferCertificateResponse
The output from the TransferCertificate operation.

TransferData
Data used to transfer a certificate to an AWS account.

UpdateCACertificateRequest
The input to the UpdateCACertificate operation.

UpdateCertificateRequest
The input for the UpdateCertificate operation.

UpdateThingRequest
The input for the UpdateThing operation.

UpdateThingResponse
The output from the UpdateThing operation.

Enums

AcceptCertificateTransferError
Errors returned by AcceptCertificateTransfer

AttachPrincipalPolicyError
Errors returned by AttachPrincipalPolicy

AttachThingPrincipalError
Errors returned by AttachThingPrincipal

CancelCertificateTransferError
Errors returned by CancelCertificateTransfer

CreateCertificateFromCsrError
Errors returned by CreateCertificateFromCsr

CreateKeysAndCertificateError
Errors returned by CreateKeysAndCertificate

CreatePolicyError
Errors returned by CreatePolicy

CreatePolicyVersionError
Errors returned by CreatePolicyVersion

CreateThingError
Errors returned by CreateThing

CreateThingTypeError
Errors returned by CreateThingType

CreateTopicRuleError
Errors returned by CreateTopicRule

DeleteCACertificateError
Errors returned by DeleteCACertificate

DeleteCertificateError
Errors returned by DeleteCertificate

DeletePolicyError
Errors returned by DeletePolicy

DeletePolicyVersionError
Errors returned by DeletePolicyVersion

DeleteRegistrationCodeError
Errors returned by DeleteRegistrationCode

DeleteThingError
Errors returned by DeleteThing

DeleteThingTypeError
Errors returned by DeleteThingType

DeleteTopicRuleError
Errors returned by DeleteTopicRule

DeprecateThingTypeError
Errors returned by DeprecateThingType

DescribeCACertificateError
Errors returned by DescribeCACertificate

DescribeCertificateError
Errors returned by DescribeCertificate

DescribeEndpointError
Errors returned by DescribeEndpoint

DescribeThingError
Errors returned by DescribeThing

DescribeThingTypeError
Errors returned by DescribeThingType

DetachPrincipalPolicyError
Errors returned by DetachPrincipalPolicy

DetachThingPrincipalError
Errors returned by DetachThingPrincipal

DisableTopicRuleError
Errors returned by DisableTopicRule

EnableTopicRuleError
Errors returned by EnableTopicRule

GetLoggingOptionsError
Errors returned by GetLoggingOptions

GetPolicyError
Errors returned by GetPolicy

GetPolicyVersionError
Errors returned by GetPolicyVersion

GetRegistrationCodeError
Errors returned by GetRegistrationCode

GetTopicRuleError
Errors returned by GetTopicRule

ListCACertificatesError
Errors returned by ListCACertificates

ListCertificatesByCAError
Errors returned by ListCertificatesByCA

ListCertificatesError
Errors returned by ListCertificates

ListOutgoingCertificatesError
Errors returned by ListOutgoingCertificates

ListPoliciesError
Errors returned by ListPolicies

ListPolicyPrincipalsError
Errors returned by ListPolicyPrincipals

ListPolicyVersionsError
Errors returned by ListPolicyVersions

ListPrincipalPoliciesError
Errors returned by ListPrincipalPolicies

ListPrincipalThingsError
Errors returned by ListPrincipalThings

ListThingPrincipalsError
Errors returned by ListThingPrincipals

ListThingTypesError
Errors returned by ListThingTypes

ListThingsError
Errors returned by ListThings

ListTopicRulesError
Errors returned by ListTopicRules

RegisterCACertificateError
Errors returned by RegisterCACertificate

RegisterCertificateError
Errors returned by RegisterCertificate

RejectCertificateTransferError
Errors returned by RejectCertificateTransfer

ReplaceTopicRuleError
Errors returned by ReplaceTopicRule

SetDefaultPolicyVersionError
Errors returned by SetDefaultPolicyVersion

SetLoggingOptionsError
Errors returned by SetLoggingOptions

TransferCertificateError
Errors returned by TransferCertificate

UpdateCACertificateError
Errors returned by UpdateCACertificate

UpdateCertificateError
Errors returned by UpdateCertificate

UpdateThingError
Errors returned by UpdateThing

Traits

Iot
Trait representing the capabilities of the AWS IoT API. AWS IoT clients implement this trait.